\name{ex01.11}
\alias{ex01.11}
\docType{data}
\title{data from exercise 1.11}
\description{
The \code{ex01.11} data frame has 40 rows and 1 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{Scores}{
      a numeric vector
    }
  }
}
%\details{}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering and the Sciences (6th ed)}, Duxbury
}
\examples{
data(ex01.11)
str(ex01.11)
}
\keyword{datasets}
