\name{ex01.12}
\alias{ex01.12}
\docType{data}
\title{data from exercise 1.12}
\description{
  Specific gravity values for various wood types.
}
\format{
  A data frame with 36 observations on the following variable.
  \describe{
    \item{SpecGrav}{numeric vector of specific gravities.}
  }
}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering and the Sciences (6th ed)}, Duxbury
}
\references{
  ``Bolted Connection Design Values Based on European Yield Model'',
  \emph{J. of Structural Engr.}, 1993: 2169-2186.
}
\examples{
str(ex01.12)
}
\keyword{datasets}
