\name{ex01.13}
\alias{ex01.13}
\docType{data}
\title{data from exercise 1.13}
\description{
  Data on tensile ultimate strength (ksi)
}
\usage{data(ex01.13)}
\format{
  A data frame with 153 observations on the following variable.
  \describe{
    \item{strength}{a numeric vector}
  }
}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering
    and the Sciences (6th ed)}, Duxbury
}
\references{
  Establishing Mechanical Property Allowables for Metals,
  \emph{J. Testing and Evaluation}, 1998: 293-299.
}
\examples{
str(ex01.13)
}
\keyword{datasets}
