\name{ex01.14}
\alias{ex01.14}
\docType{data}
\title{data from exercise 1.14}
\description{
    The \code{ex01.14} data frame has 129 rows and 1 columns.
}
\format{
    This data frame contains the following columns:
    \describe{
	\item{Rate}{
	    a numeric vector of shower-flow rates (L/min)
	}
    }
}
\details{
    The shower-flow rates for a sample of houses in Perth, Australia.
}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering
    and the Sciences (6th ed)}, Duxbury 
}
\references{    
  James, I. R. and Knuiman, M. W. (1987) An application of Bayes
  methodology to the analysis of diary records in a water use
  study. \emph{Journal of the American  Statistical Association},
  \bold{82}, 705--711.
}
\examples{
str(ex01.14)
}
\keyword{datasets}
