\name{ex01.15}
\alias{ex01.15}
\docType{data}
\title{data from exercise 1.15}
\description{
  Scores for different brands of peanut butter according to type.
}
\usage{data(ex01.15)}
\format{
  A data frame with 37 observations on the following 2 variables.
  \describe{
    \item{Score}{a numeric vector}
    \item{Type}{a factor with levels \code{Creamy} \code{Crunchy}}
  }
}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering
    and the Sciences (6th ed)}, Duxbury
}
\references{
  \emph{Consumer Reports}, Sept. 1990.
}
\examples{
str(ex01.15)
}
\keyword{datasets}
