\name{ex01.18}
\alias{ex01.18}
\docType{data}
\title{data from exercise 1.18}
\description{
  The \code{ex01.18} data frame has 18 rows and 2 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{Number.of.papers}{
      a numeric vector
    }
    \item{Frequency}{
      a numeric vector
    }
  }
}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering and the Sciences (6th ed)}, Duxbury
}
\examples{
str(ex01.18)
}
\keyword{datasets}
