\name{ex01.19}
\alias{ex01.19}
\docType{data}
\title{data from exercise 1.19}
\description{
  The \code{ex01.19} data frame has 15 rows and 2 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{Number.of.particles}{
      a numeric vector
    }
    \item{Frequency}{
      a numeric vector
    }
  }
}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering and the Sciences (6th ed)}, Duxbury
}
\examples{
str(ex01.19)
}
\keyword{datasets}
