\name{ex01.20}
\alias{ex01.20}
\docType{data}
\title{data from exercise 1.20}
\description{
  The \code{ex01.20} data frame has 47 rows and 1 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{C1}{
      a numeric vector
    }
  }
}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering and the Sciences (6th ed)}, Duxbury
}
\examples{
str(ex01.20)
}
\keyword{datasets}
