\name{ex01.21}
\alias{ex01.21}
\docType{data}
\title{data from exercise 1.21}
\description{
  The \code{ex01.21} data frame has 47 rows and 2 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{y}{
      a numeric vector
    }
    \item{z}{
      a numeric vector
    }
  }
}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering and the Sciences (6th ed)}, Duxbury
}
\examples{
str(ex01.21)
}
\keyword{datasets}
