\name{ex01.23}
\alias{ex01.23}
\docType{data}
\title{data from exercise 1.23}
\description{
  The \code{ex01.23} data frame has 100 rows and 1 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{cycles}{
      a numeric vector
    }
  }
}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering and the Sciences (6th ed)}, Duxbury
}
\examples{
str(ex01.23)
}
\keyword{datasets}
