\name{ex01.29}
\alias{ex01.29}
\docType{data}
\title{data from exercise 1.29}
\description{
  The \code{ex01.29} data frame has 60 rows and 1 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{complaint}{
      a factor with levels
      \code{B} 
      \code{C} 
      \code{F} 
      \code{J} 
      \code{M} 
      \code{N} 
      \code{O} 
    }
  }
}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering and the Sciences (6th ed)}, Duxbury
}
\examples{
str(ex01.29)
}
\keyword{datasets}
