\name{ex01.32}
\alias{ex01.32}
\docType{data}
\title{data from exercise 1.32}
\description{
  The \code{ex01.32} data frame has 14 rows and 2 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{Value}{
      a numeric vector
    }
    \item{Cumulative}{
      a numeric vector of cumulative percentages
    }
  }
}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering and the Sciences (6th ed)}, Duxbury
}
\examples{
str(ex01.32)
}
\keyword{datasets}
