\name{ex01.38}
\alias{ex01.38}
\docType{data}
\title{data from exercise 1.38}
\description{
  The \code{ex01.38} data frame has 9 rows and 1 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{pressure}{
      a numeric vector
    }
  }
}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering and the Sciences (6th ed)}, Duxbury
}
\examples{
str(ex01.38)
}
\keyword{datasets}
