\name{ex01.39}
\alias{ex01.39}
\docType{data}
\title{data from exercise 1.39}
\description{
  The \code{ex01.39} data frame has 16 rows and 1 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{lives}{
      a numeric vector
    }
  }
}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering and the Sciences (6th ed)}, Duxbury
}
\examples{
str(ex01.39)
}
\keyword{datasets}
