\name{ex01.43}
\alias{ex01.43}
\docType{data}
\title{data from exercise 1.43}
\description{
  The \code{ex01.43} data frame has 10 rows and 1 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{Lifetime}{
      a numeric vector
    }
  }
}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering and the Sciences (6th ed)}, Duxbury
}
\examples{
str(ex01.43)
}
\keyword{datasets}
