\name{ex01.50}
\alias{ex01.50}
\docType{data}
\title{data from exercise 1.50}
\description{
  Monetary awards in lawsuits.
}
\format{
  A data frame with 27 observations on the following variable.
  \describe{
    \item{awards}{a numeric vector}
  }
}
\details{
  Awards is a "normative" group of 27 cases similar to
  \emph{Genessy v. Digital Equipment Corp.}
}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering and the Sciences (6th ed)}, Duxbury
}
\examples{
str(ex01.50)
}
\keyword{datasets}
