\name{ex01.54}
\alias{ex01.54}
\docType{data}
\title{data from exercise 1.54}
\description{
  The \code{ex01.54} data frame has 11 rows and 1 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{C1}{
      a numeric vector
    }
  }
}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering and the Sciences (6th ed)}, Duxbury
}
\examples{
str(ex01.54)
}
\keyword{datasets}
