\name{ex01.60}
\alias{ex01.60}
\docType{data}
\title{data from exercise 1.60}
\description{
  The \code{ex01.60} data frame has 23 rows and 2 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{strength}{
      a numeric vector of weld strengths
    }
    \item{type}{
      a factor with levels \code{Cannister} or \code{Test} indicating the type
      of weld.
    }
  }
}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering and the Sciences (6th ed)}, Duxbury
}
\examples{
str(ex01.60)
}
\keyword{datasets}
