\name{ex01.63}
\alias{ex01.63}
\docType{data}
\title{data from exercise 1.63}
\description{
  The \code{ex01.63} data frame has 7 rows and 3 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{FlowRate.125}{
      a numeric vector
    }
    \item{FlowRate.160}{
      a numeric vector
    }
    \item{FlowRate.200}{
      a numeric vector
    }
  }
}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering and the Sciences (6th ed)}, Duxbury
}
\examples{
str(ex01.63)
}
\keyword{datasets}
