\name{ex01.65}
\alias{ex01.65}
\docType{data}
\title{data from exercise 1.65}
\description{
  The \code{ex01.65} data frame has 4 rows and 2 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{HC}{
      a numeric vector
    }
    \item{CO}{
      a numeric vector
    }
  }
}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering and the Sciences (6th ed)}, Duxbury
}
\examples{
str(ex01.65)
}
\keyword{datasets}
