\name{ex01.67}
\alias{ex01.67}
\docType{data}
\title{data from exercise 1.67}
\description{
  The \code{ex01.67} data frame has 15 rows and 1 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{CO.conc}{
      a numeric vector
    }
  }
}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering and the Sciences (6th ed)}, Duxbury
}
\examples{
str(ex01.67)
}
\keyword{datasets}
