\name{ex01.70}
\alias{ex01.70}
\docType{data}
\title{data from exercise 1.70}
\description{
  Oxygen consumption for 15 subjects on two types of exercise
}
\format{
  A data frame with 15 observations on the following 2 variables.
  \describe{
    \item{Weight}{a numeric vector of oxygen consumption when weight training}
    \item{Treadmill}{a numeric vector of oxygen consumption on a treadmill}
  }
}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering and the Sciences (6th ed)}, Duxbury
}
\references{
  ``Effect of Weight Training Exercise and Treadmill Exercise on
  Post-Exercise Oxygen Consumption'', \emph{Medicine and Science in
    Sport and Exercise}, 1998: 518--522.
}
\examples{
str(ex01.70)
}
\keyword{datasets}
