\name{ex01.72}
\alias{ex01.72}
\docType{data}
\title{data from exercise 1.72}
\description{
  Benzodiazepine receptor binding in individuals suffering from
  Posttraumatic Stress Disorder (PTSD).
}
\format{
  A data frame with 13 observations on the following 2 variables.
  \describe{
    \item{PTSD}{a numeric vector of receptor binding measures for
      individuals suffering from PTSD.}
    \item{Healthy}{a numeric vector of receptor binding measures for
      healthy individuals}
  }
}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering
    and the Sciences (6th ed)}, Duxbury
}
\references{
  ``Decreased Benzodiazepine Receptor Binding in Prefrontal Cortex in
  Combat-Related Posttraumatic Stress Disorder'', \emph{Amer. J. of
    Psychiatry}, 2000: 1120--1126.
}
\examples{
str(ex01.72)
}
\keyword{datasets}
