\name{ex01.73}
\alias{ex01.73}
\docType{data}
\title{data from exercise 1.73}
\description{
  The \code{ex01.73} data frame has 20 rows and 1 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{C1}{
      a numeric vector
    }
  }
}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering and the Sciences (6th ed)}, Duxbury
}
\examples{
str(ex01.73)
}
\keyword{datasets}
