\name{ex01.75}
\alias{ex01.75}
\docType{data}
\title{data from exercise 1.75}
\description{
  The \code{ex01.75} data frame has 15 rows and 3 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{Type.1}{
      a numeric vector
    }
    \item{Type.2}{
      a numeric vector
    }
    \item{Type.3}{
      a numeric vector
    }
  }
}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering and the Sciences (6th ed)}, Duxbury
}
\examples{
str(ex01.75)
}
\keyword{datasets}
