\name{ex01.80}
\alias{ex01.80}
\docType{data}
\title{data from exercise 1.80}
\description{
  Frequency table of the lengths (km) of bus routes.
}
\format{
  A data frame with 15 observations on the following 2 variables.
  \describe{
    \item{Length}{a factor with levels \code{10-<12} \code{12-<14} \code{14-<16} \code{16-<18} \code{18-<20} \code{20-<22} \code{22-<24} \code{24-<26} \code{26-<28} \code{28-<30} \code{30-<35} \code{35-<40} \code{40-<45} \code{6-<8} \code{8-<10}}
    \item{Frequency}{a numeric vector}
  }
}
%\details{}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering and the Sciences (6th ed)}, Duxbury
}
\references{
  Planning of City Bus Routes, \emph{J. of the Institute of Engineers},
  1995: 211-215.
}
\examples{
str(ex01.80)
summary(ex01.80)
}
\keyword{datasets}
