\name{ex01.83}
\alias{ex01.83}
\docType{data}
\title{data from exercise 1.83}
\description{
  Rainfall (acre-feet) from 26 seeded clouds.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{rainfall}{
      a numeric vector
    }
  }
}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering and the Sciences (6th ed)}, Duxbury
}
\references{
  A Bayesian Analysis of a Multiplicative Treatment Effect in Weather
  Modification, \emph{Technometrics}, 1975: 161-166.
}
\examples{
str(ex01.83)
}
\keyword{datasets}
