\name{ex04.83}
\alias{ex04.83}
\docType{data}
\title{data from exercise 4.83}
\description{
  Coating thickness for low-viscosity paint.
}
\format{
  A data frame with 16 observations on the following variable.
  \describe{
    \item{thickness}{a numeric vector}
  }
}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering
    and the Sciences (6th ed)}, Duxbury
}
\references{
  ``Achieving a Target Value for a Manufacturing Process: A Case
  Study'', \emph{J. of Quality Technology}, 1992: 22--26.
}
\examples{
str(ex04.83)
}
\keyword{datasets}
