\name{ex04.84}
\alias{ex04.84}
\docType{data}
\title{data from exercise 4.84}
\description{
  Toughness for high-strength concrete and values for Weibull plot
}
\format{
  A data frame with 18 observations on the following 2 variables.
  \describe{
    \item{obsv}{a numeric vector}
    \item{p}{a numeric vector}
  }
}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering
    and the Sciences (6th ed)}, Duxbury
}
\references{
  ``A Probabilistic Model of Fracture in Concrete and Size Effects on
  Fracture Toughness'', \emph{Magazine of Concrete Research}, 1996:
  311-320.
}
\examples{
str(ex04.84)
}
\keyword{datasets}
