\name{ex04.90}
\alias{ex04.90}
\docType{data}
\title{data from exercise 4.90}
\description{
  Sample of measurement errors
}
\format{
  A data frame with 10 observations on the following variable.
  \describe{
    \item{measerr}{a numeric vector - probably simulated.}
  }
}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering
    and the Sciences (6th ed)}, Duxbury
}
\examples{
str(ex04.90)
}
\keyword{datasets}
