\name{ex06.01}
\alias{ex06.01}
\docType{data}
\title{data from exercise 6.1}
\description{
  The \code{ex06.01} data frame has 27 rows and 1 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{C1}{
      a numeric vector
    }
  }
}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering and the Sciences (6th ed)}, Duxbury
}
\examples{
str(ex06.01)
}
\keyword{datasets}
