\name{ex06.02}
\alias{ex06.02}
\docType{data}
\title{data from exercise 6.2}
\description{
  The \code{ex06.02} data frame has 20 rows and 1 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{C1}{
      a factor with levels
      \code{C} 
      \code{H} 
      \code{S} 
      \code{T} 
    }
  }
}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering and the Sciences (6th ed)}, Duxbury
}
\examples{
str(ex06.02)
}
\keyword{datasets}
