\name{ex06.04}
\alias{ex06.04}
\docType{data}
\title{data from exercise 6.4}
\description{
  The \code{ex06.04} data frame has 20 rows and 1 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{C1}{
      a numeric vector
    }
  }
}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering and the Sciences (6th ed)}, Duxbury
}
\examples{
str(ex06.04)
}
\keyword{datasets}
