\name{ex06.05}
\alias{ex06.05}
\docType{data}
\title{data from exercise 6.5}
\description{
  The \code{ex06.05} data frame has 5 rows and 3 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{Book.value}{
      a numeric vector
    }
    \item{Audited.value}{
      a numeric vector
    }
    \item{Error}{
      a numeric vector
    }
  }
}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering and the Sciences (6th ed)}, Duxbury
}
\examples{
str(ex06.05)
}
\keyword{datasets}
