\name{ex06.09}
\alias{ex06.09}
\docType{data}
\title{data from exercise 6.9}
\description{
  The \code{ex06.09} data frame has 8 rows and 2 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{scratches}{
      a numeric vector
    }
    \item{frequency}{a numeric vector}
  }
}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering and the Sciences (6th ed)}, Duxbury
}
\examples{
str(ex06.09)
}
\keyword{datasets}
