\name{ex07.10}
\alias{ex07.10}
\docType{data}
\title{data from exercise 7.10}
\description{
  The \code{ex07.10} data frame has 15 rows and 1 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{lifetime}{
      a numeric vector
    }
  }
}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering and the Sciences (6th ed)}, Duxbury
}
\examples{
str(ex07.10)
}
\keyword{datasets}
