\name{ex07.45}
\alias{ex07.45}
\docType{data}
\title{data from exercise 7.45}
\description{
  The \code{ex07.45} data frame has 22 rows and 1 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{toughnss}{
      a numeric vector
    }
  }
}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering and the Sciences (6th ed)}, Duxbury
}
\examples{
str(ex07.45)
}
\keyword{datasets}
