\name{ex07.47}
\alias{ex07.47}
\docType{data}
\title{data from exercise 7.47}
\description{
  The \code{ex07.47} data frame has 48 rows and 1 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{strength}{
      a numeric vector
    }
  }
}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering and the Sciences (6th ed)}, Duxbury
}
\examples{
str(ex07.47)
}
\keyword{datasets}
