\name{ex07.49}
\alias{ex07.49}
\docType{data}
\title{data from exercise 7.49}
\description{
  The \code{ex07.49} data frame has 18 rows and 1 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{ResidGas}{
      a numeric vector
    }
  }
}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering and the Sciences (6th ed)}, Duxbury
}
\examples{
str(ex07.49)
}
\keyword{datasets}
