\name{ex07.56}
\alias{ex07.56}
\docType{data}
\title{data from exercise 7.56}
\description{
The \code{ex07.56} data frame has 16 rows and 1 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{pul.comp}{
a numeric vector
}
}
}
\details{
}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering and the Sciences (6th ed)}, Duxbury
}
\examples{
data(ex07.56)
}
\keyword{datasets}
