\name{ex08.32}
\alias{ex08.32}
\docType{data}
\title{data from exercise 8.32}
\description{
  The \code{ex08.32} data frame has 12 rows and 1 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{C1}{
      a numeric vector
    }
  }
}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering and the Sciences (6th ed)}, Duxbury
}
\examples{
str(ex08.32)
}
\keyword{datasets}
