\name{ex08.54}
\alias{ex08.54}
\docType{data}
\title{data from exercise 8.54}
\description{
  Amount of organic matter in soil samples.
}
\format{
  A data frame with 30 observations on the following variable.
  \describe{
    \item{percorg}{a numeric vector of percentages of organic matter}
  }
}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering and the Sciences (6th ed)}, Duxbury
}
\references{
  ``Engineering Proeprties of Soil'', \emph{Soil Science}, 1998: 93--102.
}
\examples{
str(ex08.54)
}
\keyword{datasets}
