\name{ex08.55}
\alias{ex08.55}
\docType{data}
\title{data from exercise 8.55}
\description{
  The \code{ex08.55} data frame has 13 rows and 1 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{times}{
      a numeric vector
    }
  }
}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering and the Sciences (6th ed)}, Duxbury
}
\examples{
str(ex08.55)
}
\keyword{datasets}
