\name{ex08.57}
\alias{ex08.57}
\docType{data}
\title{data from exercise 8.57}
\description{
  The \code{ex08.57} data frame has 6 rows and 1 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{CO.conc}{
      a numeric vector
    }
  }
}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering and the Sciences (6th ed)}, Duxbury
}
\examples{
str(ex08.57)
}
\keyword{datasets}
