\name{ex08.66}
\alias{ex08.66}
\docType{data}
\title{data from exercise 8.66}
\description{
  The \code{ex08.66} data frame has 8 rows and 1 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{SoilHeat}{
      a numeric vector
    }
  }
}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering and the Sciences (6th ed)}, Duxbury
}
\examples{
str(ex08.66)
}
\keyword{datasets}
