\name{ex08.80}
\alias{ex08.80}
\docType{data}
\title{data from exercise 8.80}
\description{
  The \code{ex08.80} data frame has 10 rows and 1 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{C1}{
      a numeric vector
    }
  }
}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering and the Sciences (6th ed)}, Duxbury
}
\examples{
str(ex08.80)
}
\keyword{datasets}
