\name{ex09.07}
\alias{ex09.07}
\docType{data}
\title{data from exercise 9.7}
\description{
  The \code{ex09.07} data frame has 2 rows and 4 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{Gender}{
      a factor with levels
      \code{Females} 
      \code{Males} 
    }
    \item{Sample.Size}{
      a numeric vector
    }
    \item{Sample.Mean}{
      a numeric vector
    }
    \item{Sample.Standard.Deviation}{
      a numeric vector
    }
  }
}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering and the Sciences (6th ed)}, Duxbury
}
\examples{
str(ex09.07)
}
\keyword{datasets}
