\name{ex09.12}
\alias{ex09.12}
\docType{data}
\title{data from exercise 9.12}
\description{
  The \code{ex09.12} data frame has 2 rows and 4 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{Age}{
      a numeric vector
    }
    \item{n}{
      a numeric vector
    }
    \item{Mean}{
      a numeric vector
    }
    \item{StdDev}{
      a numeric vector
    }
  }
}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering and the Sciences (6th ed)}, Duxbury
}
\examples{
str(ex09.12)
}
\keyword{datasets}
