\name{ex09.16}
\alias{ex09.16}
\docType{data}
\title{data from exercise 9.16}
\description{
  The \code{ex09.16} data frame has 2 rows and 3 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{Type}{
      a numeric vector
    }
    \item{Sample.Average}{
      a numeric vector
    }
    \item{Sample.Standard.Deviation}{
      a numeric vector
    }
  }
}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering and the Sciences (6th ed)}, Duxbury
}
\examples{
str(ex09.16)
}
\keyword{datasets}
