\name{ex09.23}
\alias{ex09.23}
\docType{data}
\title{data from exercise 9.23}
\description{
  The \code{ex09.23} data frame has 32 rows and 2 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{extens}{
      a numeric vector
    }
    \item{quality}{
      a factor with levels
      \code{H} 
      \code{P} 
    }
  }
}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering and the Sciences (6th ed)}, Duxbury
}
\examples{
str(ex09.23)
}
\keyword{datasets}
