\name{ex09.25}
\alias{ex09.25}
\docType{data}
\title{data from exercise 9.25}
\description{
  The \code{ex09.25} data frame has 2 rows and 4 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{Condition}{
      a factor with levels
      \code{LBP} 
      \code{No LBP} 
    }
    \item{Sample.size}{
      a numeric vector
    }
    \item{Sample.mean}{
      a numeric vector
    }
    \item{Sample.SD}{
      a numeric vector
    }
  }
}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering and the Sciences (6th ed)}, Duxbury
}
\examples{
str(ex09.25)
}
\keyword{datasets}
