\name{ex09.28}
\alias{ex09.28}
\docType{data}
\title{data from exercise 9.28}
\description{
  Single-step recovery from a forward fall
}
\format{
  A data frame with 15 observations on the following 2 variables.
  \describe{
    \item{Lean}{a numeric vector of the maximum lean angle from which a
      subject could still recover in one step.}
    \item{Age}{a factor with levels \code{O} \code{Y} indicating if the
      subject is an older female (67-81 years) or a younger female
      (21-29 years)}
  }
}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering and the Sciences (6th ed)}, Duxbury

}
\references{
  ``Age and Gender Differences in Single-Step Recovery from a Forward
  Fall'', \emph{J. of Gerontology}, 1999: M44-M50.
}
\examples{
str(ex09.28)
}
\keyword{datasets}
