\name{ex09.29}
\alias{ex09.29}
\docType{data}
\title{data from exercise 9.29}
\description{
  The \code{ex09.29} data frame has 2 rows and 4 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{Beverage}{
      a factor with levels
      \code{Cola} 
      \code{Strawberry drink} 
    }
    \item{Sample.size}{
      a numeric vector
    }
    \item{Sample.mean}{
      a numeric vector
    }
    \item{Sample.standard.deviation}{
      a numeric vector
    }
  }
}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering and the Sciences (6th ed)}, Duxbury
}
\examples{
str(ex09.29)
}
\keyword{datasets}
