\name{ex09.33}
\alias{ex09.33}
\docType{data}
\title{data from exercise 9.33}
\description{
  The \code{ex09.33} data frame has 2 rows and 4 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{Treatment}{
      a factor with levels
      \code{Control} 
      \code{Steroid} 
    }
    \item{Sample.size}{
      a numeric vector
    }
    \item{Sample.mean}{
      a numeric vector
    }
    \item{Sample.standard.deviation}{
      a numeric vector
    }
  }
}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering and the Sciences (6th ed)}, Duxbury
}
\examples{
str(ex09.33)
}
\keyword{datasets}
