\name{ex09.36}
\alias{ex09.36}
\docType{data}
\title{data from exercise 9.36}
\description{
  The \code{ex09.36} data frame has 8 rows and 3 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{Fabric}{
      a numeric vector
    }
    \item{U}{
      a numeric vector
    }
    \item{A}{
      a numeric vector
    }
  }
}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering and the Sciences (6th ed)}, Duxbury
}
\examples{
str(ex09.36)
}
\keyword{datasets}
