\name{ex09.37}
\alias{ex09.37}
\docType{data}
\title{data from exercise 9.37}
\description{
  The \code{ex09.37} data frame has 33 rows and 3 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{House}{
      a numeric vector
    }
    \item{Indoor}{
      a numeric vector
    }
    \item{Outdoor}{
      a numeric vector
    }
  }
}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering and the Sciences (6th ed)}, Duxbury
}
\examples{
str(ex09.37)
}
\keyword{datasets}
