\name{ex09.38}
\alias{ex09.38}
\docType{data}
\title{data from exercise 9.38}
\description{
  The \code{ex09.38} data frame has 15 rows and 3 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{Test.condition}{
      a numeric vector
    }
    \item{Normal}{
      a numeric vector
    }
    \item{High}{
      a numeric vector
    }
  }
}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering and the Sciences (6th ed)}, Duxbury
}
\examples{
str(ex09.38)
}
\keyword{datasets}
