\name{ex09.40}
\alias{ex09.40}
\docType{data}
\title{data from exercise 9.40}
\description{
  The \code{ex09.40} data frame has 16 rows and 3 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{Period}{
      a numeric vector
    }
    \item{Pipe}{
      a numeric vector
    }
    \item{Brush}{
      a numeric vector
    }
  }
}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering and the Sciences (6th ed)}, Duxbury
}
\examples{
str(ex09.40)
}
\keyword{datasets}
