\name{ex09.41}
\alias{ex09.41}
\docType{data}
\title{data from exercise 9.41}
\description{
  The \code{ex09.41} data frame has 9 rows and 3 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{Subject}{
      a numeric vector
    }
    \item{Black}{
      a numeric vector
    }
    \item{White}{
      a numeric vector
    }
  }
}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering and the Sciences (6th ed)}, Duxbury
}
\examples{
str(ex09.41)
}
\keyword{datasets}
