\name{ex09.43}
\alias{ex09.43}
\docType{data}
\title{data from exercise 9.43}
\description{
  Difference between age at onset and at diagnosis for children with childhood
  Cushing's disease.
}
\format{
  This data frame contains 15 observations on the following variable:
  \describe{
    \item{difference}{
      a numeric vector
    }
  }
}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering and the Sciences (6th ed)}, Duxbury
}
\references{
  ``Treatment of Cushing's Disease in Childhood and Adolescence by
  Transphenoidal Microadenomectomy'', \emph{New England J. of Medicine},
  1984: 889.
}
\examples{
str(ex09.43)
}
\keyword{datasets}
