\name{ex09.44}
\alias{ex09.44}
\docType{data}
\title{data from exercise 9.44}
\description{
  Modulus of elasticity of lumber samples at different times after loading
}
\usage{data(ex09.44)}
\format{
  A data frame with 16 observations on the following 2 variables.
  \describe{
    \item{X1min}{a numeric vector giving the modulus of elasticity
      1 minute after loading}
    \item{X4weeks}{a numeric vector giving the modulus of elasticity
      4 weeks after loading}
  }
}
%\details{}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering and the Sciences (6th ed)}, Duxbury
}
%\references{}
\examples{
data(ex09.44)
}
\keyword{datasets}
