\name{ex09.63}
\alias{ex09.63}
\docType{data}
\title{data from exercise 9.63}
\description{
  The \code{ex09.63} data frame has 4 rows and 2 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{Epoxy}{
      a numeric vector
    }
    \item{MMA}{
      a numeric vector
    }
  }
}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering and the Sciences (6th ed)}, Duxbury
}
\examples{
str(ex09.63)
}
\keyword{datasets}
