\name{ex09.65}
\alias{ex09.65}
\docType{data}
\title{data from exercise 9.65}
\description{
  The \code{ex09.65} data frame has 2 rows and 4 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{Method}{
      a factor with levels
      \code{Fixed} 
      \code{Floating} 
    }
    \item{n}{
      a numeric vector
    }
    \item{mean}{
      a numeric vector
    }
    \item{SD}{
      a numeric vector
    }
  }
}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering and the Sciences (6th ed)}, Duxbury
}
\examples{
str(ex09.65)
}
\keyword{datasets}
