\name{ex09.66}
\alias{ex09.66}
\docType{data}
\title{data from exercise 9.66}
\description{
  The \code{ex09.66} data frame has 8 rows and 2 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{Fertilizer.plots}{
      a numeric vector
    }
    \item{Control.plots}{
      a numeric vector
    }
  }
}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering and the Sciences (6th ed)}, Duxbury
}
\examples{
str(ex09.66)
}
\keyword{datasets}
