\name{ex09.68}
\alias{ex09.68}
\docType{data}
\title{data from exercise 9.68}
\description{
  The \code{ex09.68} data frame has 35 rows and 2 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{density}{
      a numeric vector
    }
    \item{sampling}{
      a factor with levels
      \code{block} 
      \code{pitcher} 
    }
  }
}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering and the Sciences (6th ed)}, Duxbury
}
\examples{
str(ex09.68)
}
\keyword{datasets}
