\name{ex09.70}
\alias{ex09.70}
\docType{data}
\title{data from exercise 9.70}
\description{
  The \code{ex09.70} data frame has 2 rows and 4 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{Type}{
      a factor with levels
      \code{With.side.coating} 
      \code{Without.side.coating} 
    }
    \item{size}{
      a numeric vector
    }
    \item{mean}{
      a numeric vector
    }
    \item{SD}{
      a numeric vector
    }
  }
}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering and the Sciences (6th ed)}, Duxbury
}
\examples{
str(ex09.70)
}
\keyword{datasets}
