\name{ex09.72}
\alias{ex09.72}
\docType{data}
\title{data from exercise 9.72}
\description{
  The \code{ex09.72} data frame has 17 rows and 3 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{Motor}{
      a numeric vector
    }
    \item{Commutator}{
      a numeric vector
    }
    \item{Pinion}{
      a numeric vector
    }
  }
}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering and the Sciences (6th ed)}, Duxbury
}
\examples{
str(ex09.72)
}
\keyword{datasets}
