\name{ex09.76}
\alias{ex09.76}
\docType{data}
\title{data from exercise 9.76}
\description{
  The \code{ex09.76} data frame has 2 rows and 4 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{Site}{
      a factor with levels
      \code{Clean} 
      \code{Steam.plant} 
    }
    \item{size}{
      a numeric vector
    }
    \item{mean}{
      a numeric vector
    }
    \item{SD}{
      a numeric vector
    }
  }
}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering and the Sciences (6th ed)}, Duxbury
}
\examples{
str(ex09.76)
}
\keyword{datasets}
