\name{ex09.77}
\alias{ex09.77}
\docType{data}
\title{data from exercise 9.77}
\description{
  The \code{ex09.77} data frame has 5 rows and 3 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{Twist.multiple}{
      a numeric vector
    }
    \item{Control.strength}{
      a numeric vector
    }
    \item{Heated.strength}{
      a numeric vector
    }
  }
}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering and the Sciences (6th ed)}, Duxbury
}
\examples{
str(ex09.77)
}
\keyword{datasets}
