\name{ex09.78}
\alias{ex09.78}
\docType{data}
\title{data from exercise 9.78}
\description{
  The \code{ex09.78} data frame has 2 rows and 4 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{Group}{
      a factor with levels
      \code{Elderly.men} 
      \code{Young} 
    }
    \item{size}{
      a numeric vector
    }
    \item{mean}{
      a numeric vector
    }
    \item{stdError}{
      a numeric vector
    }
  }
}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering and the Sciences (6th ed)}, Duxbury
}
\examples{
str(ex09.78)
}
\keyword{datasets}
