\name{ex09.82}
\alias{ex09.82}
\docType{data}
\title{data from exercise 9.82}
\description{
  Energy intake compared to energy expenditure for 7 soccer players
}
\format{
  A data frame with 7 observations on the following 2 variables.
  \describe{
    \item{expend}{a numeric vector}
    \item{intake}{a numeric vector}
  }
}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering
    and the Sciences (6th ed)}, Duxbury
}
\references{
  ``Measurement of Total Energy Expenditure by the Double Labelled Water
  Method in Professional Soccer Players'', \emph{J. of Sport Sciences},
  2002: 391--397.
}
\examples{
str(ex09.82)
}
\keyword{datasets}
