\name{ex09.86}
\alias{ex09.86}
\docType{data}
\title{data from exercise 9.86}
\description{
  The \code{ex09.86} data frame has 4 rows and 3 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{Treatment}{
      a numeric vector
    }
    \item{n}{
      a numeric vector
    }
    \item{SD}{
      a numeric vector
    }
  }
}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering and the Sciences (6th ed)}, Duxbury
}
\examples{
str(ex09.86)
}
\keyword{datasets}
