\name{ex09.88}
\alias{ex09.88}
\docType{data}
\title{data from exercise 9.88}
\description{
  The \code{ex09.88} data frame has 8 rows and 2 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{Carpet}{
      a numeric vector
    }
    \item{NoCarpet}{
      a numeric vector
    }
  }
}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering and the Sciences (6th ed)}, Duxbury
}
\examples{
str(ex09.88)
}
\keyword{datasets}
