\name{ex09.92}
\alias{ex09.92}
\docType{data}
\title{data from exercise 9.92}
\description{
  The \code{ex09.92} data frame has 8 rows and 3 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{Number}{
      a numeric vector
    }
    \item{Region1}{
      a numeric vector
    }
    \item{Region2}{
      a numeric vector
    }
  }
}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering and the Sciences (6th ed)}, Duxbury
}
\examples{
str(ex09.92)
}
\keyword{datasets}
