\name{ex10.06}
\alias{ex10.06}
\docType{data}
\title{data from exercise 10.6}
\description{
  The \code{ex10.06} data frame has 40 rows and 2 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{totalFe}{
      a numeric vector
    }
    \item{type}{
      a numeric vector
    }
  }
}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering and the Sciences (6th ed)}, Duxbury
}
\examples{
str(ex10.06)
}
\keyword{datasets}
