\name{ex10.08}
\alias{ex10.08}
\docType{data}
\title{data from exercise 10.8}
\description{
  The \code{ex10.08} data frame has 35 rows and 2 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{stiffnss}{
      a numeric vector
    }
    \item{length}{
      a numeric vector
    }
  }
}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering and the Sciences (6th ed)}, Duxbury
}
\examples{
str(ex10.08)
}
\keyword{datasets}
