\name{ex10.09}
\alias{ex10.09}
\docType{data}
\title{data from exercise 10.9}
\description{
  The \code{ex10.09} data frame has 6 rows and 4 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{Wheat}{
      a numeric vector
    }
    \item{Barley}{
      a numeric vector
    }
    \item{Maize}{
      a numeric vector
    }
    \item{Oats}{
      a numeric vector
    }
  }
}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering and the Sciences (6th ed)}, Duxbury
}
\examples{
str(ex10.09)
}
\keyword{datasets}
