\name{ex10.18}
\alias{ex10.18}
\docType{data}
\title{data from exercise 10.18}
\description{
  The \code{ex10.18} data frame has 20 rows and 2 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{growth}{
      a numeric vector
    }
    \item{hormone}{
      a numeric vector
    }
  }
}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering and the Sciences (6th ed)}, Duxbury
}
\examples{
str(ex10.18)
}
\keyword{datasets}
