\name{ex10.26}
\alias{ex10.26}
\docType{data}
\title{data from exercise 10.26}
\description{
  The \code{ex10.26} data frame has 26 rows and 2 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{PAPFUA}{
      a numeric vector
    }
    \item{brand}{
      a factor with levels
      \code{BlueBonnet} 
      \code{Chiffon} 
      \code{Fleischman} 
      \code{Imperial} 
      \code{Mazola} 
      \code{Parkay} 
    }
  }
}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering and the Sciences (6th ed)}, Duxbury
}
\examples{
str(ex10.26)
}
\keyword{datasets}
