\name{ex10.27}
\alias{ex10.27}
\docType{data}
\title{data from exercise 10.27}
\description{
  The \code{ex10.27} data frame has 24 rows and 2 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{Folacin}{
      a numeric vector
    }
    \item{Brand}{
      a numeric vector
    }
  }
}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering and the Sciences (6th ed)}, Duxbury
}
\examples{
str(ex10.27)
}
\keyword{datasets}
