\name{ex10.32}
\alias{ex10.32}
\docType{data}
\title{data from exercise 10.32}
\description{
  The \code{ex10.32} data frame has 5 rows and 4 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{A}{
      a numeric vector
    }
    \item{B}{
      a numeric vector
    }
    \item{C}{
      a numeric vector
    }
    \item{D}{
      a numeric vector
    }
  }
}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering and the Sciences (6th ed)}, Duxbury
}
\examples{
str(ex10.32)
}
\keyword{datasets}
