\name{ex10.41}
\alias{ex10.41}
\docType{data}
\title{data from exercise 10.41}
\description{
The \code{ex10.41} data frame has 12 rows and 2 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{Percenta}{
a numeric vector
}
\item{Lab}{
a numeric vector
}
}
}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering and the Sciences (6th ed)}, Duxbury
}
\examples{
str(ex10.41)
}
\keyword{datasets}
