\name{ex10.44}
\alias{ex10.44}
\docType{data}
\title{data from exercise 10.44}
\description{
  The \code{ex10.44} data frame has 12 rows and 2 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{Strength}{
      a numeric vector
    }
    \item{Mortar}{
      a numeric vector
    }
  }
}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering and the Sciences (6th ed)}, Duxbury
}
\examples{
str(ex10.44)
}
\keyword{datasets}
