\name{ex11.04}
\alias{ex11.04}
\docType{data}
\title{data from exercise 11.4}
\description{
  The \code{ex11.04} data frame has 12 rows and 3 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{Coverage}{
      a numeric vector
    }
    \item{Roller}{
      a numeric vector
    }
    \item{Paint}{
      a numeric vector
    }
  }
}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering and the Sciences (6th ed)}, Duxbury
}
\examples{
str(ex11.04)
}
\keyword{datasets}
