\name{ex11.05}
\alias{ex11.05}
\docType{data}
\title{data from exercise 11.5}
\description{
  The \code{ex11.05} data frame has 20 rows and 3 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{force}{
      a numeric vector
    }
    \item{angle}{
      a numeric vector
    }
    \item{connectr}{
      a numeric vector
    }
  }
}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering and the Sciences (6th ed)}, Duxbury
}
\examples{
str(ex11.05)
}
\keyword{datasets}
