\name{ex11.08}
\alias{ex11.08}
\docType{data}
\title{data from exercise 11.8}
\description{
  The \code{ex11.08} data frame has 30 rows and 3 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{epiniphr}{
      a numeric vector
    }
    \item{Anesthet}{
      a numeric vector
    }
    \item{Subject}{
      a numeric vector
    }
  }
}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering and the Sciences (6th ed)}, Duxbury
}
\examples{
str(ex11.08)
}
\keyword{datasets}
