\name{ex11.09}
\alias{ex11.09}
\docType{data}
\title{data from exercise 11.9}
\description{
  The \code{ex11.09} data frame has 36 rows and 3 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{effort}{
      a numeric vector
    }
    \item{Type}{
      a factor with levels
      \code{T1} 
      \code{T2} 
      \code{T3} 
      \code{T4} 
    }
    \item{Subject}{
      a numeric vector
    }
  }
}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering and the Sciences (6th ed)}, Duxbury
}
\examples{
str(ex11.09)
}
\keyword{datasets}
