\name{ex11.15}
\alias{ex11.15}
\docType{data}
\title{data from exercise 11.15}
\description{
  The \code{ex11.15} data frame has 18 rows and 4 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{Sand}{
      a numeric vector
    }
    \item{Carbon}{
      a numeric vector
    }
    \item{Hardness}{
      a numeric vector
    }
    \item{Strength}{
      a numeric vector
    }
  }
}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering and the Sciences (6th ed)}, Duxbury
}
\examples{
str(ex11.15)
}
\keyword{datasets}
