\name{ex11.16}
\alias{ex11.16}
\docType{data}
\title{data from exercise 11.16}
\description{
  The \code{ex11.16} data frame has 18 rows and 3 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{Response}{
      a numeric vector
    }
    \item{Formulat}{
      a numeric vector
    }
    \item{Speed}{
      a numeric vector
    }
  }
}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering and the Sciences (6th ed)}, Duxbury
}
\examples{
str(ex11.16)
}
\keyword{datasets}
