\name{ex11.17}
\alias{ex11.17}
\docType{data}
\title{data from exercise 11.17}
\description{
  The \code{ex11.17} data frame has 18 rows and 3 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{acidity}{
      a numeric vector
    }
    \item{Coal}{
      a numeric vector
    }
    \item{NaOH}{
      a numeric vector
    }
  }
}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering and the Sciences (6th ed)}, Duxbury
}
\examples{
str(ex11.17)
}
\keyword{datasets}
