\name{ex11.34}
\alias{ex11.34}
\docType{data}
\title{data from exercise 11.34}
\description{
  The \code{ex11.34} data frame has 36 rows and 4 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{Sales}{
      a numeric vector
    }
    \item{store}{
      a numeric vector
    }
    \item{week}{
      a numeric vector
    }
    \item{shelf}{
      a numeric vector
    }
  }
}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering and the Sciences (6th ed)}, Duxbury
}
\examples{
str(ex11.34)
}
\keyword{datasets}
