\name{ex11.35}
\alias{ex11.35}
\docType{data}
\title{data from exercise 11.35}
\description{
  The \code{ex11.35} data frame has 25 rows and 4 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{Moisture}{
      a numeric vector
    }
    \item{plant}{
      a numeric vector
    }
    \item{leafsize}{
      a numeric vector
    }
    \item{time}{
      a numeric vector
    }
  }
}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering and the Sciences (6th ed)}, Duxbury
}
\examples{
str(ex11.35)
}
\keyword{datasets}
