\name{ex11.40}
\alias{ex11.40}
\docType{data}
\title{data from exercise 11.40}
\description{
  The \code{ex11.40} data frame has 32 rows and 5 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{sizing}{
      a numeric vector
    }
    \item{conc}{
      a numeric vector
    }
    \item{pH}{
      a numeric vector
    }
    \item{tempture}{
      a numeric vector
    }
    \item{time}{
      a numeric vector
    }
  }
}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering and the Sciences (6th ed)}, Duxbury
}
\examples{
str(ex11.40)
}
\keyword{datasets}
