\name{ex11.42}
\alias{ex11.42}
\docType{data}
\title{data from exercise 11.42}
\description{
  The \code{ex11.42} data frame has 48 rows and 5 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{consump}{
      a numeric vector
    }
    \item{roof}{
      a numeric vector
    }
    \item{power}{
      a numeric vector
    }
    \item{scrap}{
      a numeric vector
    }
    \item{charge}{
      a numeric vector
    }
  }
}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering and the Sciences (6th ed)}, Duxbury
}
\examples{
str(ex11.42)
}
\keyword{datasets}
