\name{ex11.48}
\alias{ex11.48}
\docType{data}
\title{data from exercise 11.48}
\description{
  The \code{ex11.48} data frame has 8 rows and 5 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{thrust}{
      a numeric vector
    }
    \item{vibratn}{
      a numeric vector
    }
    \item{tempture}{
      a numeric vector
    }
    \item{altitude}{
      a numeric vector
    }
    \item{firing}{
      a numeric vector
    }
  }
}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering and the Sciences (6th ed)}, Duxbury
}
\examples{
str(ex11.48)
}
\keyword{datasets}
