\name{ex11.50}
\alias{ex11.50}
\docType{data}
\title{data from exercise 11.50}
\description{
  The \code{ex11.50} data frame has 45 rows and 3 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{Smooth}{
      a numeric vector
    }
    \item{Drying}{
      a numeric vector
    }
    \item{Fabric}{
      a numeric vector
    }
  }
}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering and the Sciences (6th ed)}, Duxbury
}
\examples{
str(ex11.50)
}
\keyword{datasets}
