\name{ex11.52}
\alias{ex11.52}
\docType{data}
\title{data from exercise 11.52}
\description{
  The \code{ex11.52} data frame has 16 rows and 3 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{clover}{
      a numeric vector
    }
    \item{plot}{
      a numeric vector
    }
    \item{rate}{
      a numeric vector
    }
  }
}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering and the Sciences (6th ed)}, Duxbury
}
\examples{
str(ex11.52)
}
\keyword{datasets}
