\name{ex11.53}
\alias{ex11.53}
\docType{data}
\title{data from exercise 11.53}
\description{
  The \code{ex11.53} data frame has 8 rows and 6 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{Run}{
      a numeric vector
    }
    \item{Spray.Volume}{
      a factor with levels
      \code{+} 
      \code{-} 
    }
    \item{Belt.Speed}{
      a factor with levels
      \code{+} 
      \code{-} 
    }
    \item{Brand}{
      a factor with levels
      \code{+} 
      \code{-} 
    }
    \item{Replication.1}{
      a numeric vector
    }
    \item{Replication.2}{
      a numeric vector
    }
  }
}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering and the Sciences (6th ed)}, Duxbury
}
\examples{
str(ex11.53)
}
\keyword{datasets}
