\name{ex11.54}
\alias{ex11.54}
\docType{data}
\title{data from exercise 11.54}
\description{
  The \code{ex11.54} data frame has 8 rows and 5 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{Sample.number}{
      a numeric vector
    }
    \item{Factor.A}{
      a numeric vector
    }
    \item{Factor.B}{
      a numeric vector
    }
    \item{Factor.C}{
      a numeric vector
    }
    \item{Resonse.EC50}{
      a numeric vector
    }
  }
}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering and the Sciences (6th ed)}, Duxbury
}
\examples{
str(ex11.54)
}
\keyword{datasets}
