\name{ex11.55}
\alias{ex11.55}
\docType{data}
\title{data from exercise 11.55}
\description{
  The \code{ex11.55} data frame has 16 rows and 1 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{Extraction}{
      a numeric vector
    }
  }
}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering and the Sciences (6th ed)}, Duxbury
}
\examples{
str(ex11.55)
}
\keyword{datasets}
