\name{ex11.56}
\alias{ex11.56}
\docType{data}
\title{data from exercise 11.56}
\description{
  The \code{ex11.56} data frame has 30 rows and 3 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{Rating}{
      a numeric vector
    }
    \item{pH}{
      a factor with levels
      \code{pH 3} 
      \code{pH 5.5} 
      \code{pH 7} 
    }
    \item{Health}{
      a factor with levels
      \code{Diseased} 
      \code{Healthy} 
    }
  }
}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering and the Sciences (6th ed)}, Duxbury
}
\examples{
str(ex11.56)
}
\keyword{datasets}
