\name{ex11.57}
\alias{ex11.57}
\docType{data}
\title{data from exercise 11.57}
\description{
  The \code{ex11.57} data frame has 54 rows and 4 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{permeability}{
      a numeric vector
    }
    \item{Pressure}{
      a numeric vector
    }
    \item{Temp}{
      a numeric vector
    }
    \item{Fabric}{
      a factor with levels
      \code{420-D} 
      \code{630-D} 
      \code{840-D} 
    }
  }
}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering and the Sciences (6th ed)}, Duxbury
}
\examples{
str(ex11.57)
}
\keyword{datasets}
