\name{ex11.61}
\alias{ex11.61}
\docType{data}
\title{data from exercise 11.61}
\description{
  The \code{ex11.61} data frame has 25 rows and 5 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{weight}{
      a numeric vector
    }
    \item{volume}{
      a numeric vector
    }
    \item{color}{
      a numeric vector
    }
    \item{size}{
      a numeric vector
    }
    \item{time}{
      a numeric vector
    }
  }
}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering and the Sciences (6th ed)}, Duxbury
}
\examples{
str(ex11.61)
}
\keyword{datasets}
