\name{ex12.01}
\alias{ex12.01}
\docType{data}
\title{data from exercise 12.1}
\description{
  The \code{ex12.01} data frame has 24 rows and 2 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{Temp}{
      a numeric vector
    }
    \item{Ratio}{
      a numeric vector
    }
  }
}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering and the Sciences (6th ed)}, Duxbury
}
\examples{
str(ex12.01)
}
\keyword{datasets}
