\name{ex12.02}
\alias{ex12.02}
\docType{data}
\title{data from exercise 12.2}
\description{
  The \code{ex12.02} data frame has 10 rows and 4 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{Engine}{
      a numeric vector
    }
    \item{Age}{
      a numeric vector
    }
    \item{Baseline}{
      a numeric vector
    }
    \item{Reformulated}{
      a numeric vector
    }
  }
}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering and the Sciences (6th ed)}, Duxbury
}
\examples{
str(ex12.02)
}
\keyword{datasets}
