\name{ex12.03}
\alias{ex12.03}
\docType{data}
\title{data from exercise 12.3}
\description{
  The \code{ex12.03} data frame has 20 rows and 2 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{x}{
      a numeric vector
    }
    \item{y}{
      a numeric vector
    }
  }
}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering and the Sciences (6th ed)}, Duxbury
}
\examples{
str(ex12.03)
}
\keyword{datasets}
