\name{ex12.13}
\alias{ex12.13}
\docType{data}
\title{data from exercise 12.13}
\description{
  Rate of deposition versus current density
}
\format{
  A data frame with 4 observations on the following 2 variables.
  \describe{
    \item{x}{a numeric vector of current densities (mA/cm-sq)}
    \item{y}{a numeric vector of deposition rates (micrometers/min)}
  }
}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering
    and the Sciences (6th ed)}, Duxbury
}
\references{
  ``Plating of 60/40 Tin/Lead Solder for Head Termination Metallurgy'',
  \emph{Plating and Surface Finishing}, Jan. 1997: 38--40
}
\examples{
str(ex12.13)
}
\keyword{datasets}
