\name{ex12.15}
\alias{ex12.15}
\docType{data}
\title{data from exercise 12.15}
\description{
  The \code{ex12.15} data frame has 27 rows and 2 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{MoE}{
      a numeric vector
    }
    \item{Strength}{
      a numeric vector
    }
  }
}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering and the Sciences (6th ed)}, Duxbury
}
\examples{
str(ex12.15)
}
\keyword{datasets}
