\name{ex12.19}
\alias{ex12.19}
\docType{data}
\title{data from exercise 12.19}
\description{
  The \code{ex12.19} data frame has 14 rows and 2 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{area}{
      a numeric vector
    }
    \item{emission}{
      a numeric vector
    }
  }
}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering and the Sciences (6th ed)}, Duxbury
}
\examples{
str(ex12.19)
}
\keyword{datasets}
