\name{ex12.20}
\alias{ex12.20}
\docType{data}
\title{data from exercise 12.20}
\description{
  The \code{ex12.20} data frame has 13 rows and 2 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{deposition}{
      a numeric vector
    }
    \item{LichenN}{
      a numeric vector
    }
  }
}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering and the Sciences (6th ed)}, Duxbury
}
\examples{
str(ex12.20)
}
\keyword{datasets}
