\name{ex12.21}
\alias{ex12.21}
\docType{data}
\title{data from exercise 12.21}
\description{
  The \code{ex12.21} data frame has 10 rows and 2 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{space}{
      a numeric vector
    }
    \item{distance}{
      a numeric vector
    }
  }
}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering and the Sciences (6th ed)}, Duxbury
}
\examples{
str(ex12.21)
}
\keyword{datasets}
