\name{ex12.24}
\alias{ex12.24}
\docType{data}
\title{data from exercise 12.24}
\description{
  The \code{ex12.24} data frame has 6 rows and 2 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{SO.2dep.}{
      a numeric vector
    }
    \item{Wt.loss}{
      a numeric vector
    }
  }
}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering and the Sciences (6th ed)}, Duxbury
}
\examples{
str(ex12.24)
}
\keyword{datasets}
