\name{ex12.36}
\alias{ex12.36}
\docType{data}
\title{data from exercise 12.36}
\description{
  Mist generation versus fluid flow from metal removal fluids
}
\format{
  A data frame with 7 observations on the following 2 variables.
  \describe{
    \item{x}{a numeric vector of fluid flow velocities for a 5\% soluble
    oil (cm/sec)}
    \item{y}{a numeric vector of the extent of mist droplets having a
      diameter less than 10 micrometers (mg/m-sq)}
  }
}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering and the Sciences (6th ed)}, Duxbury

}
\references{
  ``Variables Affecting Mist Generation from Metal Removing Fluids'',
  \emph{Lubrication Engr.}, 2002: 10--17.
}
\examples{
str(ex12.36)
}
\keyword{datasets}
