\name{ex12.37}
\alias{ex12.37}
\docType{data}
\title{data from exercise 12.37}
\description{
  The \code{ex12.37} data frame has 10 rows and 2 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{pressure}{
      a numeric vector
    }
    \item{time}{
      a numeric vector
    }
  }
}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering and the Sciences (6th ed)}, Duxbury
}
\examples{
str(ex12.37)
}
\keyword{datasets}
