\name{ex12.50}
\alias{ex12.50}
\docType{data}
\title{data from exercise 12.50}
\description{
  Relaxation time in crystals as a function of the strength of the
  external biasing magnetic field.
}
\format{
  This data frame contains 11 observations on the following variables:
  \describe{
    \item{field}{
      a numeric vector of the strength of the magnetic field (KG)
    }
    \item{time}{
      a numeric vector of the relaxation time (microseconds)
    }
  }
}
%\details{}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering
    and the Sciences (6th ed)}, Duxbury}
\references{
  ``An Optical Faraday Rotation Technique for the Determination of
  Magnetic Relaxation Times'', \emph{IEEE Trans. Magnetics}, June 1968:
  175--178.
}
\examples{
str(ex12.50)
}
\keyword{datasets}
