\name{ex12.54}
\alias{ex12.54}
\docType{data}
\title{data from exercise 12.54}
\description{
  The \code{ex12.54} data frame has 14 rows and 2 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{distance}{
      a numeric vector
    }
    \item{yield}{
      a numeric vector
    }
  }
}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering and the Sciences (6th ed)}, Duxbury
}
\examples{
str(ex12.54)
}
\keyword{datasets}
