\name{ex12.55}
\alias{ex12.55}
\docType{data}
\title{data from exercise 12.55}
\description{
  The \code{ex12.55} data frame has 12 rows and 2 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{age}{
      a numeric vector
    }
    \item{damaged}{
      a numeric vector
    }
  }
}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering and the Sciences (6th ed)}, Duxbury
}
\examples{
str(ex12.55)
}
\keyword{datasets}
