\name{ex12.58}
\alias{ex12.58}
\docType{data}
\title{data from exercise 12.58}
\description{
  The \code{ex12.58} data frame has 12 rows and 2 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{TOST}{
      a numeric vector
    }
    \item{RBOT}{
      a numeric vector
    }
  }
}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering and the Sciences (6th ed)}, Duxbury
}
\examples{
str(ex12.58)
}
\keyword{datasets}
