\name{ex12.61}
\alias{ex12.61}
\docType{data}
\title{data from exercise 12.61}
\description{
  Muscular endurance as a function of maximal lactate level.
}
\format{
  A data frame with 14 observations on the following 2 variables.
  \describe{
    \item{x}{a numeric vector of maximal lactate level}
    \item{y}{a numeric vector of muscular endurance}
  }
}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering
    and the Sciences (6th ed)}, Duxbury
}
\references{
  ``Objective Effects of a Six Months' Endurance and Strength Training
  Program in Outpatients with Congestive Heart Failure'', \emph{Medicine
    and Science in Sports and Exercise}, 1999: 1102--1107.
}
\examples{
str(ex12.61)
}
\keyword{datasets}
