\name{ex12.62}
\alias{ex12.62}
\docType{data}
\title{data from exercise 12.62}
\description{
  The \code{ex12.62} data frame has 14 rows and 2 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{x}{
      a numeric vector
    }
    \item{y}{
      a numeric vector
    }
  }
}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering and the Sciences (6th ed)}, Duxbury
}
\examples{
str(ex12.62)
}
\keyword{datasets}
