\name{ex12.65}
\alias{ex12.65}
\docType{data}
\title{data from exercise 12.65}
\description{
  The \code{ex12.65} data frame has 10 rows and 2 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{blood}{
      a numeric vector
    }
    \item{gasoline}{
      a numeric vector
    }
  }
}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering and the Sciences (6th ed)}, Duxbury
}
\examples{
str(ex12.65)
}
\keyword{datasets}
