\name{ex12.68}
\alias{ex12.68}
\docType{data}
\title{data from exercise 12.68}
\description{
  The \code{ex12.68} data frame has 8 rows and 2 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{RDF}{
      a numeric vector
    }
    \item{eff}{
      a numeric vector
    }
  }
}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering and the Sciences (6th ed)}, Duxbury
}
\examples{
str(ex12.68)
}
\keyword{datasets}
