\name{ex12.69}
\alias{ex12.69}
\docType{data}
\title{data from exercise 12.69}
\description{
  The \code{ex12.69} data frame has 13 rows and 2 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{drain.wt}{
      a numeric vector
    }
    \item{Cl.trace}{
      a numeric vector
    }
  }
}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering and the Sciences (6th ed)}, Duxbury
}
\examples{
str(ex12.69)
}
\keyword{datasets}
