\name{ex12.71}
\alias{ex12.71}
\docType{data}
\title{data from exercise 12.71}
\description{
  The \code{ex12.71} data frame has 13 rows and 2 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{austenite}{
      a numeric vector
    }
    \item{wearLoss}{
      a numeric vector
    }
  }
}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering and the Sciences (6th ed)}, Duxbury
}
\examples{
str(ex12.71)
}
\keyword{datasets}
