\name{ex12.72}
\alias{ex12.72}
\docType{data}
\title{data from exercise 12.72}
\description{
  The \code{ex12.72} data frame has 9 rows and 2 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{CO}{
      a numeric vector
    }
    \item{Noy}{
      a numeric vector
    }
  }
}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering and the Sciences (6th ed)}, Duxbury
}
\examples{
str(ex12.72)
}
\keyword{datasets}
