\name{ex12.82}
\alias{ex12.82}
\docType{data}
\title{data from exercise 12.82}
\description{
  Percentage removal of contaminants versus inlet temperature.
}
\format{
  A data frame with 33 observations on the following 2 variables.
  \describe{
    \item{temp}{a numeric vector of inlet temperatures}
    \item{removal}{a numeric vector of removal percentages}
  }
}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering
    and the Sciences (6th ed)}, Duxbury
}
\references{
  ``Treatment of Mixed Hydrogen Sulfide and Organic Vapors in Rock
  Medium Biofilter'', \emph{Water Environment Research}, 2001: 426-435.
}
\examples{
str(ex12.82)
}
\keyword{datasets}
