\name{ex12.83}
\alias{ex12.83}
\docType{data}
\title{data from exercise 12.83}
\description{
  Blood glucose level in fish versus time since stress applied.
}
\format{
  A data frame with 24 observations on the following 2 variables.
  \describe{
    \item{time}{a numeric vector of the time since stress was applied}
    \item{bloodgluc}{a numeric vector of blood glucose levels}
  }
}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering
    and the Sciences (6th ed)}, Duxbury
}
\references{
  ``Evaluation of Simple Instruments for the Measurement of Blood
  Glucose and Lactate, and Plasma Protein as Stress Indicators in
  Fish'', \emph{J. of the World Aquaculture Society}, 1999: 276--284.
}
\examples{
str(ex12.83)
}
\keyword{datasets}
