\name{ex12.84}
\alias{ex12.84}
\docType{data}
\title{data from exercise 12.84}
\description{
  Air displacement versus hydrostatic weighing for measuring body fat
}
\format{
  A data frame with 20 observations on the following 2 variables.
  \describe{
    \item{HW}{a numeric vector of results from hydrostatic weighing}
    \item{BOD}{a numeric vector of results from an air displacement device}
  }
}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering
    and the Sciences (6th ed)}, Duxbury
}
\references{
  ``Evaluating the BOD POD for Assessing Body Fat in Collegiate Football
  Players'', \emph{Medicine and Science in Sports and Exercise}, 1999:
  1350--1356.
}
\examples{
str(ex12.84)
}
\keyword{datasets}
