\name{ex13.02}
\alias{ex13.02}
\docType{data}
\title{data from exercise 13.2}
\description{
  The \code{ex13.02} data frame has 9 rows and 2 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{rate}{
      a numeric vector
    }
    \item{stdResid}{
      a numeric vector
    }
  }
}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering and the Sciences (6th ed)}, Duxbury
}
\examples{
str(ex13.02)
}
\keyword{datasets}
