\name{ex13.04}
\alias{ex13.04}
\docType{data}
\title{data from exercise 13.4}
\description{
  The \code{ex13.04} data frame has 10 rows and 2 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{thicknss}{
      a numeric vector
    }
    \item{current}{
      a numeric vector
    }
  }
}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering and the Sciences (6th ed)}, Duxbury
}
\examples{
str(ex13.04)
}
\keyword{datasets}
