\name{ex13.05}
\alias{ex13.05}
\docType{data}
\title{data from exercise 13.5}
\description{
  Ice thickness versus elapsed time
}
\format{
  A data frame with 33 observations on the following 2 variables.
  \describe{
    \item{time}{a numeric vector of elapsed time (hr)}
    \item{icethick}{a numeric vector of ice thickness (mm)}
  }
}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering
    and the Sciences (6th ed)}, Duxbury} 
\references{
  ``Laborator Study of Anchor Ice Growth'', \emph{J. of Cold Regions
    Engr.}, 2001: 60-66.
}
\examples{
str(ex13.05)
}
\keyword{datasets}
