\name{ex13.06}
\alias{ex13.06}
\docType{data}
\title{data from exercise 13.6}
\description{
  The \code{ex13.06} data frame has 6 rows and 2 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{density}{
      a numeric vector
    }
    \item{moisture}{
      a numeric vector
    }
  }
}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering and the Sciences (6th ed)}, Duxbury
}
\examples{
str(ex13.06)
}
\keyword{datasets}
