\name{ex13.07}
\alias{ex13.07}
\docType{data}
\title{data from exercise 13.7}
\description{
  The \code{ex13.07} data frame has 5 rows and 2 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{exposure}{
      a numeric vector
    }
    \item{weight}{
      a numeric vector
    }
  }
}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering and the Sciences (6th ed)}, Duxbury
}
\examples{
str(ex13.07)
}
\keyword{datasets}
