\name{ex13.08}
\alias{ex13.08}
\docType{data}
\title{data from exercise 13.8}
\description{
  Relationship between heart rate and oxygen uptake during exercise
}
\format{
  A data frame with 15 observations on the following 2 variables.
  \describe{
    \item{HR}{a numeric vector of heart rates (as a percentage of maximum
    rate)}
    \item{VO2}{a numeric vector of oxygen uptake (as a percentage of
      maximum uptake)}
  }
}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering
    and the Sciences (6th ed)}, Duxbury
}
\references{
  ``The Relationship between Heart Rate and Oxygen Uptake During
  Non-Steady State Exercise'', \emph{Ergonomics}, 2000: 1578-1592.
}
\examples{
str(ex13.08)
}
\keyword{datasets}
