\name{ex13.09}
\alias{ex13.09}
\docType{data}
\title{data from exercise 13.9}
\description{
  The \code{ex13.09} data frame has 44 rows and 3 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{x}{
      a numeric vector
    }
    \item{y}{
      a numeric vector
    }
    \item{set}{
      a factor with levels
      \code{a} 
      \code{b} 
      \code{c} 
      \code{d} 
    }
  }
}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering and the Sciences (6th ed)}, Duxbury
}
\examples{
str(ex13.09)
}
\keyword{datasets}
