\name{ex13.14}
\alias{ex13.14}
\docType{data}
\title{data from exercise 13.14}
\description{
  The \code{ex13.14} data frame has 14 rows and 2 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{weight}{
      a numeric vector
    }
    \item{clearnce}{
      a numeric vector
    }
  }
}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering and the Sciences (6th ed)}, Duxbury
}
\examples{
str(ex13.14)
}
\keyword{datasets}
