\name{ex13.16}
\alias{ex13.16}
\docType{data}
\title{data from exercise 13.16}
\description{
  The \code{ex13.16} data frame has 12 rows and 2 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{Spectral}{
      a numeric vector
    }
    \item{ln.L178.}{
      a numeric vector
    }
  }
}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering and the Sciences (6th ed)}, Duxbury
}
\examples{
str(ex13.16)
}
\keyword{datasets}
