\name{ex13.18}
\alias{ex13.18}
\docType{data}
\title{data from exercise 13.18}
\description{
  The \code{ex13.18} data frame has 4 rows and 2 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{Conc.}{
      a numeric vector
    }
    \item{pH}{
      a numeric vector
    }
  }
}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering and the Sciences (6th ed)}, Duxbury
}
\examples{
str(ex13.18)
}
\keyword{datasets}
