\name{ex13.21}
\alias{ex13.21}
\docType{data}
\title{data from exercise 13.21}
\description{
  The \code{ex13.21} data frame has 8 rows and 2 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{thicknss}{
      a numeric vector
    }
    \item{conduct}{
      a numeric vector
    }
  }
}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering and the Sciences (6th ed)}, Duxbury
}
\examples{
str(ex13.21)
}
\keyword{datasets}
