\name{ex13.25}
\alias{ex13.25}
\docType{data}
\title{data from exercise 13.25}
\description{
  The \code{ex13.25} data frame has 14 rows and 2 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{Success}{
      a numeric vector
    }
    \item{Failure}{
      a numeric vector
    }
  }
}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering and the Sciences (6th ed)}, Duxbury
}
\examples{
data(ex13.25)
}
\keyword{datasets}
