\name{ex13.27}
\alias{ex13.27}
\docType{data}
\title{data from exercise 13.27}
\description{
  The \code{ex13.27} data frame has 8 rows and 2 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{time}{
      a numeric vector
    }
    \item{conc}{
      a numeric vector
    }
  }
}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering and the Sciences (6th ed)}, Duxbury
}
\examples{
str(ex13.27)
}
\keyword{datasets}
