\name{ex13.29}
\alias{ex13.29}
\docType{data}
\title{data from exercise 13.29}
\description{
  The \code{ex13.29} data frame has 5 rows and 2 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{x}{
      a numeric vector
    }
    \item{y}{
      a numeric vector
    }
  }
}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering and the Sciences (6th ed)}, Duxbury
}
\examples{
str(ex13.29)
}
\keyword{datasets}
