\name{ex13.30}
\alias{ex13.30}
\docType{data}
\title{data from exercise 13.30}
\description{
  The \code{ex13.30} data frame has 14 rows and 2 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{water}{
      a numeric vector
    }
    \item{yield}{
      a numeric vector
    }
  }
}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering and the Sciences (6th ed)}, Duxbury
}
\examples{
str(ex13.30)
}
\keyword{datasets}
