\name{ex13.33}
\alias{ex13.33}
\docType{data}
\title{data from exercise 13.33}
\description{
  The \code{ex13.33} data frame has 7 rows and 2 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{angle}{
      a numeric vector
    }
    \item{efficncy}{
      a numeric vector
    }
  }
}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering and the Sciences (6th ed)}, Duxbury
}
\examples{
str(ex13.33)
}
\keyword{datasets}
