\name{ex13.34}
\alias{ex13.34}
\docType{data}
\title{data from exercise 13.34}
\description{
  The \code{ex13.34} data frame has 13 rows and 2 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{cations}{
      a numeric vector
    }
    \item{exchange}{
      a numeric vector
    }
  }
}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering and the Sciences (6th ed)}, Duxbury
}
\examples{
str(ex13.34)
}
\keyword{datasets}
