\name{ex13.35}
\alias{ex13.35}
\docType{data}
\title{data from exercise 13.35}
\description{
  The \code{ex13.35} data frame has 5 rows and 2 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{tempture}{
      a numeric vector
    }
    \item{rate}{
      a numeric vector
    }
  }
}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering and the Sciences (6th ed)}, Duxbury
}
\examples{
str(ex13.35)
}
\keyword{datasets}
