\name{ex13.49}
\alias{ex13.49}
\docType{data}
\title{data from exercise 13.49}
\description{
  The \code{ex13.49} data frame has 12 rows and 3 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{yield}{
      a numeric vector
    }
    \item{temp}{
      a numeric vector
    }
    \item{sunshine}{
      a numeric vector
    }
  }
}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering and the Sciences (6th ed)}, Duxbury
}
\examples{
str(ex13.49)
}
\keyword{datasets}
