\name{ex13.51}
\alias{ex13.51}
\docType{data}
\title{data from exercise 13.51}
\description{
  The \code{ex13.51} data frame has 14 rows and 3 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{shear}{
      a numeric vector
    }
    \item{depth}{
      a numeric vector
    }
    \item{water}{
      a numeric vector
    }
  }
}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering and the Sciences (6th ed)}, Duxbury
}
\examples{
str(ex13.51)
}
\keyword{datasets}
