\name{ex13.52}
\alias{ex13.52}
\docType{data}
\title{data from exercise 13.52}
\description{
  Production of beta-carotene according to amounts of lineolic acid,
  kerosene, and antioxidant.
}
\format{
  A data frame with 20 observations on the following 4 variables.
  \describe{
    \item{Linoleic}{a numeric vector}
    \item{Kerosene}{a numeric vector}
    \item{Antiox}{a numeric vector}
    \item{Betacaro}{a numeric vector}
  }
}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering
    and the Sciences (6th ed)}, Duxbury
}
\references{
  ``Optimization of the Production of beta-Carotene from Molasses by
  Blakeslea Trispora'', \emph{J. of Chemical Technology and
    Biotechnology}, 2002: 933-943.
}
\examples{
str(ex13.52)
}
\keyword{datasets}
