\name{ex13.53}
\alias{ex13.53}
\docType{data}
\title{data from exercise 13.53}
\description{
  Deposition of atmospheric pollutants in snowpacks.
}
\format{
  A data frame with 17 observations on the following 3 variables.
  \describe{
    \item{x1}{a numeric vector of a derived predictor variable}
    \item{x2}{a numeric vector of a second derived predictor variable}
    \item{filth}{a numeric vector of the amount of deposition over a
      specified time period}
  }
}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering
    and the Sciences (6th ed)}, Duxbury
} 
\references{
  ``Atmospheric PAH Deposition: Deposition Velocities and Washout
  Ratios'', \emph{J. of Environmental Engineering}, 2002: 186-195.
}
\examples{
str(ex13.53)
}
\keyword{datasets}
