\name{ex13.54}
\alias{ex13.54}
\docType{data}
\title{data from exercise 13.54}
\description{
  The \code{ex13.54} data frame has 31 rows and 5 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{Bright}{
      a numeric vector
    }
    \item{H2O2}{
      a numeric vector
    }
    \item{NaOH}{
      a numeric vector
    }
    \item{Silicate}{
      a numeric vector
    }
    \item{Tempture}{
      a numeric vector
    }
  }
}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering and the Sciences (6th ed)}, Duxbury
}
\examples{
str(ex13.54)
}
\keyword{datasets}
