\name{ex13.55}
\alias{ex13.55}
\docType{data}
\title{data from exercise 13.55}
\description{
  The \code{ex13.55} data frame has 10 rows and 3 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{q}{
      a numeric vector
    }
    \item{a}{
      a numeric vector
    }
    \item{b}{
      a numeric vector
    }
  }
}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering and the Sciences (6th ed)}, Duxbury
}
\examples{
str(ex13.55)
}
\keyword{datasets}
