\name{ex13.68}
\alias{ex13.68}
\docType{data}
\title{data from exercise 13.68}
\description{
  The \code{ex13.68} data frame has 16 rows and 2 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{Log.edges.}{
      a numeric vector
    }
    \item{Log.time.}{
      a numeric vector
    }
  }
}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering and the Sciences (6th ed)}, Duxbury
}
\examples{
str(ex13.68)
}
\keyword{datasets}
