\name{ex13.69}
\alias{ex13.69}
\docType{data}
\title{data from exercise 13.69}
\description{
  The \code{ex13.69} data frame has 18 rows and 2 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{Pressure}{
      a numeric vector
    }
    \item{Temperature}{
      a numeric vector
    }
  }
}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering and the Sciences (6th ed)}, Duxbury
}
\examples{
str(ex13.69)
}
\keyword{datasets}
