\name{ex13.71}
\alias{ex13.71}
\docType{data}
\title{data from exercise 13.71}
\description{
  The \code{ex13.71} data frame has 32 rows and 7 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{Obs}{
      a numeric vector
    }
    \item{pdconc}{
      a numeric vector
    }
    \item{niconc}{
      a numeric vector
    }
    \item{pH}{
      a numeric vector
    }
    \item{temp}{
      a numeric vector
    }
    \item{currdens}{
      a numeric vector
    }
    \item{pallcont}{
      a numeric vector
    }
  }
}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering
    and the Sciences (6th ed)}, Duxbury
}
\examples{
str(ex13.71)
}
\keyword{datasets}
