\name{ex13.73}
\alias{ex13.73}
\docType{data}
\title{data from exercise 13.73}
\description{
  The \code{ex13.73} data frame has 8 rows and 2 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{power}{
      a numeric vector
    }
    \item{freq}{
      a numeric vector
    }
  }
}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering and the Sciences (6th ed)}, Duxbury
}
\examples{
str(ex13.73)
}
\keyword{datasets}
