\name{ex13.74}
\alias{ex13.74}
\docType{data}
\title{data from exercise 13.74}
\description{
  The \code{ex13.74} data frame has 12 rows and 2 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{log.con.}{
      a numeric vector
    }
    \item{Li2O}{
      a numeric vector
    }
  }
}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering and the Sciences (6th ed)}, Duxbury
}
\examples{
str(ex13.74)
}
\keyword{datasets}
