\name{ex13.75}
\alias{ex13.75}
\docType{data}
\title{data from exercise 13.75}
\description{
  The \code{ex13.75} data frame has 10 rows and 2 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{height}{
      a numeric vector
    }
    \item{log.Mn.}{
      a numeric vector
    }
  }
}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering and the Sciences (6th ed)}, Duxbury
}
\examples{
str(ex13.75)
}
\keyword{datasets}
