\name{ex13.76}
\alias{ex13.76}
\docType{data}
\title{data from exercise 13.76}
\description{
  The \code{ex13.76} data frame has 9 rows and 3 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{x1}{
      a numeric vector
    }
    \item{x2}{
      a numeric vector
    }
    \item{y}{
      a numeric vector
    }
  }
}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering and the Sciences (6th ed)}, Duxbury
}
\examples{
str(ex13.76)
}
\keyword{datasets}
