\name{ex14.09}
\alias{ex14.09}
\docType{data}
\title{data from exercise 14.9}
\description{
  The \code{ex14.09} data frame has 40 rows and 1 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{time}{
      a numeric vector
    }
  }
}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering and the Sciences (6th ed)}, Duxbury
}
\examples{
str(ex14.09)
}
\keyword{datasets}
