\name{ex14.11}
\alias{ex14.11}
\docType{data}
\title{data from exercise 14.11}
\description{
  The \code{ex14.11} data frame has 45 rows and 1 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{diam}{
      a numeric vector
    }
  }
}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering and the Sciences (6th ed)}, Duxbury
}
\examples{
str(ex14.11)
}
\keyword{datasets}
