\name{ex14.13}
\alias{ex14.13}
\docType{data}
\title{data from exercise 14.13}
\description{
  The \code{ex14.13} data frame has 3 rows and 2 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{ovaries.developed}{
      a numeric vector
    }
    \item{Observed.count}{
      a numeric vector
    }
  }
}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering and the Sciences (6th ed)}, Duxbury
}
\examples{
str(ex14.13)
}
\keyword{datasets}
