\name{ex14.17}
\alias{ex14.17}
\docType{data}
\title{data from exercise 14.17}
\description{
  The \code{ex14.17} data frame has 13 rows and 2 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{borers}{
      a numeric vector
    }
    \item{freq}{
      a numeric vector
    }
  }
}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering and the Sciences (6th ed)}, Duxbury
}
\examples{
str(ex14.17)
}
\keyword{datasets}
