\name{ex14.18}
\alias{ex14.18}
\docType{data}
\title{data from exercise 14.18}
\description{
  The \code{ex14.18} data frame has 5 rows and 2 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{Rate}{
      a factor with levels
      \code{.100-below .150} 
      \code{.150-below .200} 
      \code{.200-below .250} 
      \code{.250 or more} 
      \code{Below .100} 
    }
    \item{Frequency}{
      a numeric vector
    }
  }
}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering and the Sciences (6th ed)}, Duxbury
}
\examples{
str(ex14.18)
}
\keyword{datasets}
