\name{ex14.20}
\alias{ex14.20}
\docType{data}
\title{data from exercise 14.20}
\description{
  The \code{ex14.20} data frame has 23 rows and 1 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{C1}{
      a numeric vector
    }
  }
}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering and the Sciences (6th ed)}, Duxbury
}
\examples{
str(ex14.20)
}
\keyword{datasets}
