\name{ex14.21}
\alias{ex14.21}
\docType{data}
\title{data from exercise 14.21}
\description{
  The \code{ex14.21} data frame has 24 rows and 1 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{C1}{
      a numeric vector
    }
  }
}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering and the Sciences (6th ed)}, Duxbury
}
\examples{
str(ex14.21)
}
\keyword{datasets}
