\name{ex14.23}
\alias{ex14.23}
\docType{data}
\title{data from exercise 14.23}
\description{
  The \code{ex14.23} data frame has 30 rows and 1 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{Strength}{
      a numeric vector
    }
  }
}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering and the Sciences (6th ed)}, Duxbury
}
\examples{
str(ex14.23)
}
\keyword{datasets}
