\name{ex14.26}
\alias{ex14.26}
\docType{data}
\title{data from exercise 14.26}
\description{
  The \code{ex14.26} data frame has 5 rows and 3 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{Treatment}{
      a factor with levels
      \code{Control} 
      \code{Eight leaves removed} 
      \code{Four leaves removed} 
      \code{Six leaves removed} 
      \code{Two leaves removed} 
    }
    \item{Matured}{
      a numeric vector
    }
    \item{Aborted}{
      a numeric vector
    }
  }
}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering and the Sciences (6th ed)}, Duxbury
}
\examples{
str(ex14.26)
}
\keyword{datasets}
