\name{ex14.27}
\alias{ex14.27}
\docType{data}
\title{data from exercise 14.27}
\description{
  The \code{ex14.27} data frame has 2 rows and 5 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{C1}{
      a factor with levels
      \code{Men} 
      \code{Women} 
    }
    \item{L.R}{
      a numeric vector
    }
    \item{L.R.1}{
      a numeric vector
    }
    \item{L.R.2}{
      a numeric vector
    }
    \item{Sample.size}{
      a numeric vector
    }
  }
}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering and the Sciences (6th ed)}, Duxbury
}
\examples{
str(ex14.27)
}
\keyword{datasets}
