\name{ex14.28}
\alias{ex14.28}
\docType{data}
\title{data from exercise 14.28}
\description{
  The \code{ex14.28} data frame has 4 rows and 4 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{Thienyla}{
      a numeric vector
    }
    \item{Solvent}{
      a numeric vector
    }
    \item{Sham}{
      a numeric vector
    }
    \item{Unhandle}{
      a numeric vector
    }
  }
}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering and the Sciences (6th ed)}, Duxbury
}
\examples{
str(ex14.28)
}
\keyword{datasets}
