\name{ex14.29}
\alias{ex14.29}
\docType{data}
\title{data from exercise 14.29}
\description{
  The \code{ex14.29} data frame has 6 rows and 3 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{M.M}{
      a numeric vector
    }
    \item{M.F}{
      a numeric vector
    }
    \item{F.F}{
      a numeric vector
    }
  }
}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering and the Sciences (6th ed)}, Duxbury
}
\examples{
str(ex14.29)
}
\keyword{datasets}
