\name{ex14.30}
\alias{ex14.30}
\docType{data}
\title{data from exercise 14.30}
\description{
  The \code{ex14.30} data frame has 12 rows and 3 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{count}{
      a numeric vector
    }
    \item{Config}{
      a numeric vector
    }
    \item{Mode}{
      a numeric vector
    }
  }
}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering and the Sciences (6th ed)}, Duxbury
}
\examples{
str(ex14.30)
}
\keyword{datasets}
