\name{ex14.38}
\alias{ex14.38}
\docType{data}
\title{data from exercise 14.38}
\description{
  The \code{ex14.38} data frame has 3 rows and 2 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{Parsitiz}{
      a numeric vector
    }
    \item{Nonparas}{
      a numeric vector
    }
  }
}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering and the Sciences (6th ed)}, Duxbury
}
\examples{
str(ex14.38)
}
\keyword{datasets}
