\name{ex14.41}
\alias{ex14.41}
\docType{data}
\title{data from exercise 14.41}
\description{
  The \code{ex14.41} data frame has 3 rows and 3 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{Never}{
      a numeric vector
    }
    \item{Occasion}{
      a numeric vector
    }
    \item{Regular}{
      a numeric vector
    }
  }
}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering and the Sciences (6th ed)}, Duxbury
}
\examples{
str(ex14.41)
}
\keyword{datasets}
