\name{ex14.42}
\alias{ex14.42}
\docType{data}
\title{data from exercise 14.42}
\description{
  The \code{ex14.42} data frame has 4 rows and 3 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{Home}{
      a numeric vector
    }
    \item{Acute}{
      a numeric vector
    }
    \item{Chronic}{
      a numeric vector
    }
  }
}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering and the Sciences (6th ed)}, Duxbury
}
\examples{
str(ex14.42)
}
\keyword{datasets}
