\name{ex15.01}
\alias{ex15.01}
\docType{data}
\title{data from exercise 15.1}
\description{
  The \code{ex15.01} data frame has 12 rows and 1 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{C1}{
      a numeric vector
    }
  }
}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering and the Sciences (6th ed)}, Duxbury
}
\examples{
str(ex15.01)
}
\keyword{datasets}
