\name{ex15.03}
\alias{ex15.03}
\docType{data}
\title{data from exercise 15.3}
\description{
  The \code{ex15.03} data frame has 14 rows and 1 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{pH}{
      a numeric vector
    }
  }
}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering and the Sciences (6th ed)}, Duxbury
}
\examples{
str(ex15.03)
}
\keyword{datasets}
