\name{ex15.05}
\alias{ex15.05}
\docType{data}
\title{data from exercise 15.5}
\description{
  The \code{ex15.05} data frame has 12 rows and 3 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{Sample}{
      a numeric vector
    }
    \item{Gravimetric}{
      a numeric vector
    }
    \item{Spectrophotometric}{
      a numeric vector
    }
  }
}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering and the Sciences (6th ed)}, Duxbury
}
\examples{
str(ex15.05)
}
\keyword{datasets}
