\name{ex15.10}
\alias{ex15.10}
\docType{data}
\title{data from exercise 15.10}
\description{
  The \code{ex15.10} data frame has 5 rows and 2 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{adhesv.1}{
      a numeric vector
    }
    \item{adhesv.2}{
      a numeric vector
    }
  }
}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering and the Sciences (6th ed)}, Duxbury
}
\examples{
str(ex15.10)
}
\keyword{datasets}
