\name{ex15.11}
\alias{ex15.11}
\docType{data}
\title{data from exercise 15.11}
\description{
  The \code{ex15.11} data frame has 14 rows and 2 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{Time}{
      a numeric vector
    }
    \item{wood}{
      a factor with levels
      \code{Oak} 
      \code{Pine} 
    }
  }
}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering and the Sciences (6th ed)}, Duxbury
}
\examples{
str(ex15.11)
}
\keyword{datasets}
