\name{ex15.12}
\alias{ex15.12}
\docType{data}
\title{data from exercise 15.12}
\description{
  The \code{ex15.12} data frame has 8 rows and 2 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{Original}{
      a numeric vector
    }
    \item{Modified}{
      a numeric vector
    }
  }
}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering and the Sciences (6th ed)}, Duxbury
}
\examples{
str(ex15.12)
}
\keyword{datasets}
