\name{ex15.13}
\alias{ex15.13}
\docType{data}
\title{data from exercise 15.13}
\description{
  The \code{ex15.13} data frame has 10 rows and 2 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{Orange.J}{
      a numeric vector
    }
    \item{Ascorbic}{
      a numeric vector
    }
  }
}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering and the Sciences (6th ed)}, Duxbury
}
\examples{
str(ex15.13)
}
\keyword{datasets}
