\name{ex15.15}
\alias{ex15.15}
\docType{data}
\title{data from exercise 15.15}
\description{
  The \code{ex15.15} data frame has 15 rows and 2 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{conc}{
      a numeric vector
    }
    \item{exposed}{
      a factor with levels
      \code{N} 
      \code{Y} 
    }
  }
}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering and the Sciences (6th ed)}, Duxbury
}
\examples{
str(ex15.15)
}
\keyword{datasets}
