\name{ex15.23}
\alias{ex15.23}
\docType{data}
\title{data from exercise 15.23}
\description{
  The \code{ex15.23} data frame has 5 rows and 4 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{Region.1}{
      a numeric vector
    }
    \item{Region.2}{
      a numeric vector
    }
    \item{Region.3}{
      a numeric vector
    }
    \item{Region.4}{
      a numeric vector
    }
  }
}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering and the Sciences (6th ed)}, Duxbury
}
\examples{
str(ex15.23)
}
\keyword{datasets}
