\name{ex15.24}
\alias{ex15.24}
\docType{data}
\title{data from exercise 15.24}
\description{
  The \code{ex15.24} data frame has 9 rows and 4 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{fasting}{
      a numeric vector
    }
    \item{pct23}{
      a numeric vector
    }
    \item{pct32}{
      a numeric vector
    }
    \item{pct67}{
      a numeric vector
    }
  }
}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering and the Sciences (6th ed)}, Duxbury
}
\examples{
str(ex15.24)
}
\keyword{datasets}
