\name{ex15.26}
\alias{ex15.26}
\docType{data}
\title{data from exercise 15.26}
\description{
  The \code{ex15.26} data frame has 10 rows and 5 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{Blocks}{
      a numeric vector
    }
    \item{A}{
      a numeric vector
    }
    \item{B}{
      a numeric vector
    }
    \item{C}{
      a numeric vector
    }
    \item{D}{
      a numeric vector
    }
  }
}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering and the Sciences (6th ed)}, Duxbury
}
\examples{
str(ex15.26)
}
\keyword{datasets}
