\name{ex15.28}
\alias{ex15.28}
\docType{data}
\title{data from exercise 15.28}
\description{
  The \code{ex15.28} data frame has 8 rows and 3 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{Subject}{
      a numeric vector
    }
    \item{Potato}{
      a numeric vector
    }
    \item{Rice}{
      a numeric vector
    }
  }
}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering and the Sciences (6th ed)}, Duxbury
}
\examples{
str(ex15.28)
}
\keyword{datasets}
