\name{ex15.29}
\alias{ex15.29}
\docType{data}
\title{data from exercise 15.29}
\description{
  The \code{ex15.29} data frame has 9 rows and 4 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{X1973}{
      a numeric vector
    }
    \item{X1974}{
      a numeric vector
    }
    \item{X1975}{
      a numeric vector
    }
    \item{X1976}{
      a numeric vector
    }
  }
}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering and the Sciences (6th ed)}, Duxbury
}
\examples{
str(ex15.29)
}
\keyword{datasets}
