\name{ex15.30}
\alias{ex15.30}
\docType{data}
\title{data from exercise 15.30}
\description{
  The \code{ex15.30} data frame has 5 rows and 4 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{Treatment.I}{
      a numeric vector
    }
    \item{Treatment.II}{
      a numeric vector
    }
    \item{Treatment.III}{
      a numeric vector
    }
    \item{Treatment.IV}{
      a numeric vector
    }
  }
}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering and the Sciences (6th ed)}, Duxbury
}
\examples{
str(ex15.30)
}
\keyword{datasets}
