\name{ex15.33}
\alias{ex15.33}
\docType{data}
\title{data from exercise 15.33}
\description{
  The \code{ex15.33} data frame has 20 rows and 1 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{C1}{
      a numeric vector
    }
  }
}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering and the Sciences (6th ed)}, Duxbury
}
\examples{
str(ex15.33)
}
\keyword{datasets}
