\name{ex16.06}
\alias{ex16.06}
\docType{data}
\title{data from exercise 16.6}
\description{
  The \code{ex16.06} data frame has 22 rows and 5 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{Obs.1}{
      a numeric vector
    }
    \item{Obs.2}{
      a numeric vector
    }
    \item{Obs.3}{
      a numeric vector
    }
    \item{Obs.4}{
      a numeric vector
    }
    \item{Obs.5}{
      a numeric vector
    }
  }
}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering
    and the Sciences (6th ed)}, Duxbury
}
\examples{
str(ex16.06)
}
\keyword{datasets}
