\name{ex16.09}
\alias{ex16.09}
\docType{data}
\title{data from exercise 16.9}
\description{
  The \code{ex16.09} data frame has 24 rows and 2 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{xbar}{
      a numeric vector
    }
    \item{stderr}{
      a numeric vector
    }
  }
}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering and the Sciences (6th ed)}, Duxbury
}
\examples{
str(ex16.09)
}
\keyword{datasets}
