\name{ex16.25}
\alias{ex16.25}
\docType{data}
\title{data from exercise 16.25}
\description{
  The \code{ex16.25} data frame has 21 rows and 3 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{C1}{
      a factor with levels
      \code{1} 
      \code{10} 
      \code{11} 
      \code{12} 
      \code{13} 
      \code{14} 
      \code{15} 
      \code{16} 
      \code{17} 
      \code{18} 
      \code{19} 
      \code{2} 
      \code{20} 
      \code{3} 
      \code{4} 
      \code{5} 
      \code{6} 
      \code{7} 
      \code{8} 
      \code{9} 
      \code{Panel} 
    }
    \item{C2}{
      a factor with levels
      \code{0.6} 
      \code{0.8} 
      \code{1} 
      \code{Area Examined} 
    }
    \item{C3}{
      a factor with levels
      \code{# Blemishes} 
      \code{1} 
      \code{10} 
      \code{12} 
      \code{2} 
      \code{3} 
      \code{4} 
      \code{5} 
      \code{6} 
    }
  }
}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering and the Sciences (6th ed)}, Duxbury
}
\examples{
str(ex16.25)
}
\keyword{datasets}
