\name{ex16.41}
\alias{ex16.41}
\docType{data}
\title{data from exercise 16.41}
\description{
  The \code{ex16.41} data frame has 22 rows and 3 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{tensile1}{
      a numeric vector
    }
    \item{tensile2}{
      a numeric vector
    }
    \item{tensile3}{
      a numeric vector
    }
  }
}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering and the Sciences (6th ed)}, Duxbury
}
\examples{
str(ex16.41)
}
\keyword{datasets}
