\name{ex16.43}
\alias{ex16.43}
\docType{data}
\title{data from exercise 16.43}
\description{
  The \code{ex16.43} data frame has 20 rows and 3 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{n}{
      a numeric vector
    }
    \item{xbar}{
      a numeric vector
    }
    \item{stderr}{
      a numeric vector
    }
  }
}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering and the Sciences (6th ed)}, Duxbury
}
\examples{
str(ex16.43)
}
\keyword{datasets}
