\name{xmp01.13}
\alias{xmp01.13}
\docType{data}
\title{data from Example 1.13}
\description{
  Crack lengths in corrosion tests
}
\format{
  A data frame with 21 observations on the following variable.
  \describe{
    \item{crackLength}{a numeric vector}
  }
}
%\details{}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering
      and the Sciences (6th ed)}, Duxbury
}
\references{
  Based on data shown in ``On the Role of Phosphorus in the Caustic
  Stress Corrosion Cracking of Low Alloy Steels'', \emph{Corrosion
    Science}, 1989: 53--68.
}
\examples{
str(xmp01.13)
}
\keyword{datasets}
