\name{xmp01.14}
\alias{xmp01.14}
\docType{data}
\title{data from Example 1.14}
\description{
  Concentrations of transferrin receptor in a sample of pregnant women.
}
\format{
  A data frame with 12 observations on the following variable.
  \describe{
    \item{concentration}{a numeric vector}
  }
}
%\details{}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering
      and the Sciences (6th ed)}, Duxbury
}
\references{
  Based on data shown in ``Serum Transferrin Receptor for the Detection
  of Iron Deficiency in Pregnancy'', \emph{Amer. J. of Clinical
    Nutrition}, 1991: 1077--1081.
}
\examples{
str(xmp01.14)
}
\keyword{datasets}
