\name{xmp01.16}
\alias{xmp01.16}
\docType{data}
\title{data from Example 1.16}
\description{
  The \code{xmp01.16} data frame has 11 rows and 1 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{Strength}{
      a numeric vector
    }
  }
}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering and the Sciences (6th ed)}, Duxbury
}
\examples{
str(xmp01.16)
}
\keyword{datasets}
