\name{xmp01.18}
\alias{xmp01.18}
\docType{data}
\title{data from Example 1.18}
\description{
    The \code{xmp01.18} data frame has 19 rows and 1 columns.
}
\format{
    This data frame contains the following columns:
    \describe{
	\item{depth}{
	    a numeric vector
	}
    }
}
\source{
    Devore, J. L. (2003) \emph{Probability and Statistics for
	Engineering and the Sciences (6th ed)}, Duxbury
}
\examples{
str(xmp01.18)
}
\keyword{datasets}
