\name{xmp01.19}
\alias{xmp01.19}
\docType{data}
\title{data from Example 1.19}
\description{
  The \code{xmp01.19} data frame has 25 rows and 1 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{C1}{
      a numeric vector
    }
  }
}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering and the Sciences (6th ed)}, Duxbury
}
\examples{
str(xmp01.19)
}
\keyword{datasets}
