\name{xmp09.04}
\alias{xmp09.04}
\docType{data}
\title{data from Example 9.4}
\description{
  The \code{xmp09.04} data frame has 2 rows and 4 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{Type}{
      a factor with levels
      \code{Graded} 
      \code{No-fines} 
    }
    \item{Sample.Size}{
      a numeric vector
    }
    \item{Sample.Average.Conductivity}{
      a numeric vector
    }
    \item{Sample.Standard.Deviation}{
      a numeric vector
    }
  }
}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering
    and the Sciences (6th ed)}, Duxbury
}
\examples{
str(xmp09.04)
}
\keyword{datasets}
