\name{xmp09.06}
\alias{xmp09.06}
\docType{data}
\title{data from Example 9.6}
\description{
  The \code{xmp09.06} data frame has 2 rows and 4 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{Fabric.Type}{
      a factor with levels
      \code{Cotton} 
      \code{Triacetate} 
    }
    \item{Sample.Size}{
      a numeric vector
    }
    \item{Sample.Mean}{
      a numeric vector
    }
    \item{Sample.Standard.Deviation}{
      a numeric vector
    }
  }
}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering
    and the Sciences (6th ed)}, Duxbury
}
\examples{
str(xmp09.06)
}
\keyword{datasets}
