\name{xmp09.07}
\alias{xmp09.07}
\docType{data}
\title{data from Example 9.7}
\description{
  Data on the tensile strength of liner specimens, both when a certain
  fusion process was used and when this process was not used.
}
\format{
  A data frame with 18 observations on the following 2 variables.
  \describe{
    \item{\code{strength}}{tensile strength (psi)}
    \item{\code{type}}{a factor with levels \code{fused} \code{nofusion}}
  }
}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering
    and the Sciences (6th ed)}, Duxbury
}
\references{
  "Effect of welding on high-density polyethylene liner", \emph{J. of
    Materials in Civil Engr.}, 1996: 94--100.
}
\examples{
str(xmp09.07)
}
\keyword{datasets}
