\name{xmp12.12}
\alias{xmp12.12}
\docType{data}
\title{data from Example 12.12}
\description{
  The \code{xmp12.12} data frame has 18 rows and 2 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{x}{
      a numeric vector
    }
    \item{y}{
      a numeric vector
    }
  }
}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering and the Sciences (6th ed)}, Duxbury
}
\examples{
str(xmp12.12)
}
\keyword{datasets}
