\name{xmp12.13}
\alias{xmp12.13}
\docType{data}
\title{data from Example 12.13}
\description{
  Corrosion of steel reinforcing bars
}
\format{
  A data frame with 18 observations on the following 2 variables.
  \describe{
    \item{x}{a numeric vector of carbonation depths (mm)}
    \item{y}{a numeric vector of strengths (MPa) for a sample of core
      specimens}
  }
}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering
    and the Sciences (6th ed)}, Duxbury
}
\references{
  ``The Carbonation of Concrete Structures in the Tropical Environment
  of Singapore'', \emph{Magazine of Concrete Research}, 1996: 293--300
}
\examples{
str(xmp12.13)
}
\keyword{datasets}
