\name{xmp12.15}
\alias{xmp12.15}
\docType{data}
\title{data from Example 12.15}
\description{
  The \code{xmp12.15} data frame has 16 rows and 2 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{ozone}{
      a numeric vector
    }
    \item{carbon}{
      a numeric vector
    }
  }
}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering and the Sciences (6th ed)}, Duxbury
}
\examples{
str(xmp12.15)
}
\keyword{datasets}
