\name{xmp13.10}
\alias{xmp13.10}
\docType{data}
\title{data from Example 13.10}
\description{
  Ultimate shear strength of a rubber compound versus cure temperature.
}
\format{
  A data frame with 8 observations on the following 3 variables.
  \describe{
    \item{x}{a numeric vector of cure temperatures (degrees Fahrenheit)}
    \item{x.}{a numeric vector of centered x values}
    \item{y}{a numeric vector of ultimate shear strength (psi)}
  }
}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering
    and the Sciences (6th ed)}, Duxbury
}
\references{
  ``A Method of Improving the Accuracy of Polynomial Regression
  Analysis'', \emph{J. Quality Technology}, 1971: 149-155.
}
\examples{
str(xmp13.10)
}
\keyword{datasets}
