\name{xmp13.15}
\alias{xmp13.15}
\docType{data}
\title{data from Example 13.15}
\description{
  The \code{xmp13.15} data frame has 13 rows and 3 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{index}{
      a numeric vector
    }
    \item{iron}{
      a numeric vector
    }
    \item{aluminum}{
      a numeric vector
    }
  }
}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering and the Sciences (6th ed)}, Duxbury
}
\examples{
str(xmp13.15)
}
\keyword{datasets}
