\name{xmp13.16}
\alias{xmp13.16}
\docType{data}
\title{data from Example 13.16}
\description{
  The \code{xmp13.16} data frame has 30 rows and 5 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{press}{
      a numeric vector
    }
    \item{HCHO}{
      a numeric vector
    }
    \item{catalyst}{
      a numeric vector
    }
    \item{temp}{
      a numeric vector
    }
    \item{time}{
      a numeric vector
    }
  }
}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering and the Sciences (6th ed)}, Duxbury
}
\examples{
str(xmp13.16)
}
\keyword{datasets}
