\name{xmp13.18}
\alias{xmp13.18}
\docType{data}
\title{data from Example 13.18}
\description{
  The \code{xmp13.18} data frame has 27 rows and 3 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{life}{
      a numeric vector
    }
    \item{speed}{
      a numeric vector
    }
    \item{load}{
      a numeric vector
    }
  }
}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering and the Sciences (6th ed)}, Duxbury
}
\examples{
str(xmp13.18)
}
\keyword{datasets}
