\name{xmp13.19}
\alias{xmp13.19}
\docType{data}
\title{data from Example 13.19}
\description{
  The \code{xmp13.19} data frame has 31 rows and 3 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{tar}{
      a numeric vector
    }
    \item{speed}{
      a numeric vector
    }
    \item{tempture}{
      a numeric vector
    }
  }
}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering and the Sciences (6th ed)}, Duxbury
}
\examples{
str(xmp13.19)
}
\keyword{datasets}
