\name{xmp13.22}
\alias{xmp13.22}
\docType{data}
\title{data from Example 13.22}
\description{
  The \code{xmp13.22} data frame has 10 rows and 3 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{Strength}{
      a numeric vector
    }
    \item{Sp.grav.}{
      a numeric vector
    }
    \item{Moisture}{
      a numeric vector
    }
  }
}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering and the Sciences (6th ed)}, Duxbury
}
\examples{
str(xmp13.22)
}
\keyword{datasets}
