\name{xmp14.03}
\alias{xmp14.03}
\docType{data}
\title{data from Example 14.3}
\description{
  The \code{xmp14.03} data frame has 24 rows and 1 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{onset}{
      a numeric vector
    }
  }
}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering and the Sciences (6th ed)}, Duxbury
}
\examples{
str(xmp14.03)
}
\keyword{datasets}
