\name{xmp14.14}
\alias{xmp14.14}
\docType{data}
\title{data from Example 14.14}
\description{
  The \code{xmp14.14} data frame has 3 rows and 3 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{Substand}{
      a numeric vector
    }
    \item{Standard}{
      a numeric vector
    }
    \item{Modern}{
      a numeric vector
    }
  }
}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering and the Sciences (6th ed)}, Duxbury
}
\examples{
str(xmp14.14)
}
\keyword{datasets}
