\name{xmp15.01}
\alias{xmp15.01}
\docType{data}
\title{data from Example 15.1}
\description{
  The \code{xmp15.01} data frame has 15 rows and 1 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{C1}{
      a numeric vector
    }
  }
}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering and the Sciences (6th ed)}, Duxbury
}
\examples{
str(xmp15.01)
}
\keyword{datasets}
