\name{xmp15.04}
\alias{xmp15.04}
\docType{data}
\title{data from Example 15.4}
\description{
  The \code{xmp15.04} data frame has 12 rows and 2 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{conc}{
      a numeric vector
    }
    \item{Area}{
      a factor with levels
      \code{Polluted} 
      \code{Unpolluted} 
    }
  }
}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering and the Sciences (6th ed)}, Duxbury
}
\examples{
str(xmp15.04)
}
\keyword{datasets}
