\name{xmp15.06}
\alias{xmp15.06}
\docType{data}
\title{data from Example 15.6}
\description{
  The \code{xmp15.06} data frame has 28 rows and 1 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{metabolc}{
      a numeric vector
    }
  }
}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering and the Sciences (6th ed)}, Duxbury
}
\examples{
str(xmp15.06)
}
\keyword{datasets}
