\name{xmp15.08}
\alias{xmp15.08}
\docType{data}
\title{data from Example 15.8}
\description{
  The \code{xmp15.08} data frame has 11 rows and 2 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{strength}{
      a numeric vector
    }
    \item{type}{
      a factor with levels
      \code{Epoxy} 
      \code{Other} 
    }
  }
}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering and the Sciences (6th ed)}, Duxbury
}
\examples{
str(xmp15.08)
}
\keyword{datasets}
