\name{xmp15.10}
\alias{xmp15.10}
\docType{data}
\title{data from Example 15.10}
\description{
  The \code{xmp15.10} data frame has 32 rows and 3 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{potental}{
      a numeric vector
    }
    \item{emotion}{
      a numeric vector
    }
    \item{subject}{
      a numeric vector
    }
  }
}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering and the Sciences (6th ed)}, Duxbury
}
\examples{
str(xmp15.10)
}
\keyword{datasets}
