\name{xmp16.04}
\alias{xmp16.04}
\docType{data}
\title{data from Example 16.4}
\description{
  The \code{xmp16.04} data frame has 22 rows and 4 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{resist1}{
      a numeric vector
    }
    \item{resist2}{
      a numeric vector
    }
    \item{resist3}{
      a numeric vector
    }
    \item{resist4}{
      a numeric vector
    }
  }
}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering and the Sciences (6th ed)}, Duxbury
}
\examples{
str(xmp16.04)
}
\keyword{datasets}
