\name{xmp16.06}
\alias{xmp16.06}
\docType{data}
\title{data from Example 16.6}
\description{
  The \code{xmp16.06} data frame has 25 rows and 1 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{defects}{
      a numeric vector
    }
  }
}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering and the Sciences (6th ed)}, Duxbury
}
\examples{
str(xmp16.06)
}
\keyword{datasets}
