\name{xmp16.07}
\alias{xmp16.07}
\docType{data}
\title{data from Example 16.7}
\description{
  The \code{xmp16.07} data frame has 24 rows and 1 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{flaws}{
      a numeric vector
    }
  }
}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering and the Sciences (6th ed)}, Duxbury
}
\examples{
str(xmp16.07)
}
\keyword{datasets}
