\name{xmp16.08}
\alias{xmp16.08}
\docType{data}
\title{data from Example 16.8}
\description{
  The \code{xmp16.08} data frame has 16 rows and 4 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{obs1}{
      a numeric vector
    }
    \item{obs2}{
      a numeric vector
    }
    \item{obs3}{
      a numeric vector
    }
    \item{obs4}{
      a numeric vector
    }
  }
}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering and the Sciences (6th ed)}, Duxbury
}
\examples{
str(xmp16.08)
}
\keyword{datasets}
