\name{xmp16.09}
\alias{xmp16.09}
\docType{data}
\title{data from Example 16.9}
\description{
  The \code{xmp16.09} data frame has 16 rows and 6 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{Sample}{
      a numeric vector
    }
    \item{xwl}{
      a numeric vector
    }
    \item{xwl...40.15}{
      a numeric vector
    }
    \item{dl}{
      a numeric vector
    }
    \item{xwl...39.85}{
      a numeric vector
    }
    \item{el}{
      a numeric vector
    }
  }
}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering
    and the Sciences (6th ed)}, Duxbury
}
\examples{
str(xmp16.09)
}
\keyword{datasets}
