\name{ex01.27}
\alias{ex01.27}
\docType{data}
\title{data from exercise 1.27}
\description{
  Number of cycles of strain to breakage for 100 yarn samples.
}
\usage{data(ex01.27)}
\format{
  A data frame with 50 observations on the following variable.
  \describe{
    \item{lifetime}{a numeric vector}
  }
}
%\details{}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering and the Sciences (6th ed)}, Duxbury
}
\references{
  \emph{Technometrics}, 1982: 63.
}
\examples{
data(ex01.27)
str(ex01.27)
}
\keyword{datasets}
