\name{ex01.77}
\alias{ex01.77}
\docType{data}
\title{data from exercise 1.77}
\description{
  The \code{ex01.77} data frame has 46 rows and 1 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{Time}{
      a numeric vector of active repair times (hours) for airborne
      communications receivers.
    }
  }
}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering and the Sciences (6th ed)}, Duxbury
}
\examples{
str(ex01.77)
}
\keyword{datasets}
