\name{ex04.91}
\alias{ex04.91}
\docType{data}
\title{data from exercise 4.91}
\description{
  Failure times for accelerated life testing of integrated circuits.
}
\usage{data(ex04.91)}
\format{
  A data frame with 16 observations on the following variable.
  \describe{
    \item{failtime}{a numeric vector of failure times (1000's of hours)}
  }
}
%\details{}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering
    and the Sciences (6th ed)}, Duxbury
}
%\references{}
\examples{
data(ex04.91)
}
\keyword{datasets}
