\name{ex07.33}
\alias{ex07.33}
\docType{data}
\title{data from exercise 7.33}
\description{
  The \code{ex07.33} data frame has 17 rows and 1 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{C1}{
      a numeric vector
    }
  }
}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering and the Sciences (6th ed)}, Duxbury
}
\examples{
str(ex07.33)
}
\keyword{datasets}
