\name{ex07.37}
\alias{ex07.37}
\docType{data}
\title{data from exercise 7.37}
\description{
The \code{ex07.37} data frame has 20 rows and 1 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{C1}{
a numeric vector
}
}
}
\details{
}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering and the Sciences (6th ed)}, Duxbury
}
\examples{
data(ex07.37)
}
\keyword{datasets}
