\name{ex07.46}
\alias{ex07.46}
\docType{data}
\title{data from exercise 7.46}
\description{
The \code{ex07.46} data frame has 15 rows and 1 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{C1}{
a numeric vector
}
}
}
\details{
}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering and the Sciences (6th ed)}, Duxbury
}
\examples{
data(ex07.46)
}
\keyword{datasets}
