\name{ex08.70}
\alias{ex08.70}
\docType{data}
\title{data from exercise 8.70}
\description{
The \code{ex08.70} data frame has 20 rows and 1 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{time}{
a numeric vector
}
}
}
%\details{}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering and the Sciences (6th ed)}, Duxbury
}
\examples{
data(ex08.70)
}
\keyword{datasets}
