\name{ex14.14}
\alias{ex14.14}
\docType{data}
\title{data from exercise 14.14}
\description{
  The \code{ex14.14} data frame has 12 rows and 2 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{x}{
      a numeric vector
    }
    \item{observed}{
      a numeric vector
    }
  }
}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering and the Sciences (6th ed)}, Duxbury
}
\examples{
str(ex14.14)
}
\keyword{datasets}
