\name{ex14.22}
\alias{ex14.22}
\docType{data}
\title{data from exercise 14.22}
\description{
  The \code{ex14.22} data frame has 25 rows and 1 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{toughnss}{
      a numeric vector
    }
  }
}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering and the Sciences (6th ed)}, Duxbury
}
\examples{
str(ex14.22)
}
\keyword{datasets}
