\name{xmp14.10}
\alias{xmp14.10}
\docType{data}
\title{data from Example 14.10}
\description{
  The \code{xmp14.10} data frame has 49 rows and 1 columns.
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{Cholstrl}{
      a numeric vector
    }
  }
}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering and the Sciences (6th ed)}, Duxbury
}
\examples{
str(xmp14.10)
}
\keyword{datasets}
